% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_flow.R
\name{biomass_flow}
\alias{biomass_flow}
\title{Calculate the consumed biomass from predator x to prey Y.}
\usage{
biomass_flow(dir = getwd(), nc_prod, nc_gen, dietcheck, prm_biol, prm_run,
  bps, fgs, bboxes, plot_diet = FALSE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc_prod}{Character string giving the filename of the productivity netcdf
file. Usually "output[...]PROD.nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_prod. In addition set dir to 'NULL' in this
case.}

\item{nc_gen}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{dietcheck}{Character string of the DietCheck.txt file. Usually
'output[...]DietCheck.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as dietcheck. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{plot_diet}{Logical indicating if you want to apply the \code{plot_diet} function or not.
Default is \code{FALSE}.}
}
\value{
Dataframe
}
\description{
Consumption data is extracted from output[...]PROD.nc. Age based groups
are stored as "Eat_" non age based groups as "Grazing_". Units are mg N m^-3 d^-1.
Factors are species, time, box and agecl (if present). We will refer to species as
pred from heron to indicate the predator perspective.
Diet constribution data is extracted from DietCheck.txt. Currently this only works
with models based on the trunk code. Units are % diet contribution (Will check if
it sums to 1 per predator/agecl combination). Factors are pred, time, agecl, prey.
The biomass flows are calculated as follows:
- Calculate consumed biomass as Eat (or Grazing) * boxvolume per time, pred, agecl, box.
- Convert to biomass in [t].
- Sum consumed biomass over the model domain per pred, time, agecl.
- Combine with diet contributions and calculate consumed biomass of prey species.
- Sum up consumed biomass per pred, time, prey.
}
\examples{
dir <- "c:/backup_z/Atlantis_models/Runs/dummy_02_ATLANTIS_NS/"
nc_prod <- "outputNorthSeaPROD.nc"
nc_gen <- "outputNorthSea.nc"
dietcheck <- "outputNorthSeaDietCheck.txt"
prm_biol <- "NorthSea_biol_fishing.prm"
prm_run <- "NorthSea_run_fishing_F.prm"
fgs <- "functionalGroups.csv"
}

