% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-biomass-spatial.R
\name{calculate_biomass_spatial}
\alias{calculate_biomass_spatial}
\title{Calculate spatiallly explicit biomass (in [t]) for each group and ageclass per timestep.}
\usage{
calculate_biomass_spatial(dir = getwd(), nc_gen, prm_biol, bps, fgs, bboxes)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc_gen}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}
}
\description{
Calculate spatiallly explicit biomass (in [t]) for each group and ageclass per timestep.
}
\examples{
\dontrun{
dir <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
nc_gen <- "outputSETAS.nc"
prm_biol <- "VMPA_setas_biol_fishing_Trunk.prm",
   prm_run = "VMPA_setas_run_fishing_F_New.prm",
   bps = load_bps(dir = d, fgs = "SETasGroups.csv", init = "init_vmpa_setas_25032013.nc"),
   fgs = "SETasGroups.csv",
   select_groups = c("Planktiv_S_Fish", "Pisciv_S_Fish", "Cephalopod", "Diatom", "Zoo"),
   bboxes = get_boundary(boxinfo = load_box(dir = d, bgm = "VMPA_setas.bgm")),
   check_acronyms = TRUE,
   modelstart = "1991-01-01",
   out = "preprocess.Rda",
   save_to_disc = FALSE,
   version_flag = 1)
   }
}

