% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-consumed-biomass.R
\name{calculate_consumed_biomass}
\alias{calculate_consumed_biomass}
\title{Calculate the consumed biomass in [t] of prey j by predator i.}
\usage{
calculate_consumed_biomass(dir = getwd(), nc_prod, nc_gen, dietcheck,
  prm_biol, prm_run, bps, fgs, bboxes)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc_prod}{Character string giving the filename of the productivity netcdf
file. Usually "output[...]PROD.nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_prod. In addition set dir to 'NULL' in this
case.}

\item{nc_gen}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{dietcheck}{Character string of the DietCheck.txt file. Usually
'output[...]DietCheck.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as dietcheck. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}
}
\value{
Dataframe with columns 'pred', 'agecl', 'polygon', 'time', 'prey'.
Consumed biomass in [t] is stored in column 'atoutput'.
}
\description{
Consumption data is extracted from output[...]PROD.nc. Age based groups
are stored as "Eat_" non age based groups as "Grazing_". Units are mg N m^-3 d^-1.
Factors are species, time, box and agecl (if present). We will refer to species as
pred from here on to indicate the predator perspective.
Diet constribution data is extracted from DietCheck.txt. Currently this only works
for models based on the trunk code. Units are % diet contribution. Factors are pred, time, agecl, prey.
The consumed biomass is calculated as follows:
- Calculate consumed biomass as Eat (or Grazing) * boxvolume per time, pred, agecl, box.
- Convert to biomass in [t].
- Combine with diet contributions and calculate consumed biomass of prey species.
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
nc_prod <- "outputNorthSeaPROD.nc"
nc_gen <- "outputNorthSea.nc"
dietcheck <- "outputNorthSeaDietCheck.txt"
prm_biol <- "NorthSea_biol_fishing.prm"
prm_run <- "NorthSea_run_fishing_F.prm"
fgs <- "functionalGroups.csv"
bps <- load_bps(dir = dir, init = "init_simple_NorthSea.nc", fgs = fgs)
bboxes <- get_boundary(load_box(dir = dir, bgm = "NorthSea.bgm"))

df <- calculate_consumed_biomass(dir, nc_prod, nc_gen, dietcheck,
                                 prm_biol, prm_run, bps, fgs, bboxes)
}

