% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-spatial-overlap.R
\name{calculate_spatial_overlap}
\alias{calculate_spatial_overlap}
\title{Calculate 3d overlap of predator groups with their prey over time using Schoener Index.}
\usage{
calculate_spatial_overlap(biomass_spatial, dietmatrix, agemat)
}
\arguments{
\item{biomass_spatial}{Biomass timeseries of each group and ageclass per polygon
and layer. This dataframe should be generated with \code{\link{calculate_biomass_spatial}}.}

\item{dietmatrix}{Availability matrix given in the biological parameter file.
This dataframe should be generated with \code{\link{load_dietmatrix}}. Please
use \code{convert_names = TRUE} in \code{load_dietmatrix}.}

\item{agemat}{First mature age class for age structured groups. This dataframe should
be generated with \code\link{prm_to_df} using "age_mat" as parameter.}
}
\value{
Schoener's similarity index ranging from 1 (perfect overlap) to
0 (zero overlap).
}
\description{
Calculate 3d overlap of predator groups with their prey over time using Schoener Index.
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
nc_gen <- "outputNorthSea.nc"
prm_biol <- "NorthSea_biol_fishing.prm"
prm_run = "NorthSea_run_fishing_F.prm"
bps = load_bps(dir, fgs = "functionalGroups.csv", init = "init_simple_NorthSea.nc")
fgs = "functionalGroups.csv"
bboxes = get_boundary(boxinfo = load_box(dir, bgm = "NorthSea.bgm"))

biomass_spatial <- calculate_biomass_spatial(dir, nc_gen, prm_biol, prm_run, bps, fgs, bboxes)
dietmatrix <- load_dietmatrix(dir, prm_biol, fgs, convert_names = TRUE)
agemat <- prm_to_df(dir = dir, prm_biol = prm_biol, fgs = fgs, group = get_age_acronyms(dir = dir, fgs = fgs),
                    parameter = "age_mat")

sp_overlap <- calculate_spatial_overlap(biomass_spatial, dietmatrix)
}

