% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-prm-cohort.R
\name{change_prm_cohort}
\alias{change_prm_cohort}
\title{Change biological parameterfile for parameters which expect multiple values.}
\usage{
change_prm_cohort(dir = getwd(), prm_biol, select_acronyms, roc, parameter,
  relative = TRUE, save_to_disc = TRUE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{select_acronyms}{Character vector of funtional groups which shall be read in.
Names have to match the ones used in the *.prm file. Check column "Code" in
"functionalGroups.csv" for clarification.}

\item{roc}{Matrix of multiplication factors which shall be applied to the old set of parameters.
Please supply one row per selected group. Each row should have as many entries as the
parameter itself. E.g. if you want to change the clearance rate for two fish groups you
need to supply a matrix with 2 rows and 10 columns.}

\item{parameter}{Character value of the model parameter which shall be changed.
Only one parameter can be selected per function call.}

\item{relative}{Logical if TRUE values are changed relative to base values. If FALSE new values can
be passed directly.}

\item{save_to_disc}{Logical indicating if the resulting prm file should be overwritten
(\code{TRUE}) or not (\code{FALSE}).}
}
\value{
parameterfile *.prm file with the new parameter values.
}
\description{
This functionis is used to help automate the calibration routine for ATLANTIS models.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
change_prm_cohort(dir = d,
   prm_biol = "VMPA_setas_biol_fishing_New.prm",
   select_acronyms = c("FPS", "FVS"),
   roc = c(2,3),
   parameter = "C",
   save_to_disc = FALSE)
}

