% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-df-names.R
\name{check_df_names}
\alias{check_df_names}
\title{Function to check the names of a dataframe.}
\usage{
check_df_names(data, expect, optional = NULL)
}
\arguments{
\item{data}{Dataframe to check.}

\item{expect}{Character vector giving the names of the expected columns}

\item{optional}{Character vector giving the names of optional columns.
Default is \code{NULL}.}
}
\description{
This function is used in most plotting routines to check if the correct dataframe
is passed.
}
\examples{
\dontrun{
check_df_names(preprocess$biomass_age, expect = c("time", "species", "atoutput", "ages"))
}
}
