% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-ages.R
\name{combine_ages}
\alias{combine_ages}
\title{Combine ageclasses to juvenile and adult stanza according to age at maturity.}
\usage{
combine_ages(data, grp_col, agemat, value_col = "atoutput")
}
\arguments{
\item{data}{Dataframe with ageclass specific information.}

\item{grp_col}{Character string giving the name of the group column in \code{data}.
E.g. 'species', 'pred', 'prey' etc.}

\item{agemat}{First mature age class for age structured groups. This dataframe should
be generated with \code\link{prm_to_df} using "age_mat" as parameter.}

\item{value_col}{Character string giving the name of the column to sum.
Default is \code{"atoutput"}.}
}
\value{
Dataframe with ageclasses combined to stanzas.
}
\description{
Combine ageclasses to juvenile and adult stanza according to age at maturity.
}
\examples{

dir <- system.file("extdata", "gns", package = "atlantistools")
nc_gen <- "outputNorthSea.nc"
prm_biol <- "NorthSea_biol_fishing.prm"
prm_run = "NorthSea_run_fishing_F.prm"
bps = load_bps(dir, fgs = "functionalGroups.csv", init = "init_simple_NorthSea.nc")
fgs = "functionalGroups.csv"
bboxes = get_boundary(boxinfo = load_box(dir, bgm = "NorthSea.bgm"))

data <- calculate_biomass_spatial(dir, nc_gen, prm_biol, prm_run, bps, fgs, bboxes)#'
agemat <- prm_to_df(dir = dir, prm_biol = prm_biol, fgs = fgs, group = get_age_acronyms(dir = dir, fgs = fgs),
                    parameter = "age_mat")

combine_ages(data, grp_col = "species", agemat = agemat)
}

