% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-groups.R
\name{combine_groups}
\alias{combine_groups}
\title{Combine values from different groups if specific groups only have a low contribution to the overall value.}
\usage{
combine_groups(data, group_col, groups, combine_thresh = 0.03)
}
\arguments{
\item{data}{Dataframe whose grouops shall be combined.}

\item{group_col}{Character string giving the name of the group column in `data`.}

\item{groups}{Vector of character strings giving the grouping variables.}

\item{combine_thresh}{Numeric indicating minimum amount to the stomach contribution.
Each prey item with a lower contribution as this treshold is assigned to the
grpup "Rest". This is necessary for species with a wide variety of food items
in their diet. Otherwise the screen will be cluttered with colors in the
dietplots. Default is 0.03.}
}
\value{
dataframe with groupgs combined to "Rest" if contribution is low.
}
\description{
Combine values from different groups if specific groups only have a low contribution to the overall value.
}
\examples{
combine_groups(preprocess_setas$diet,
   group_col = "prey",
   groups = c("time", "pred", "habitat"),
   combine_thresh = 0.03)
}

