% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-bgm.R
\name{convert_bgm}
\alias{convert_bgm}
\title{Transform data from bgm-file to map dataframe.}
\usage{
convert_bgm(dir = getwd(), bgm)
}
\arguments{
\item{dir}{Path of the Atlantis model folder.}

\item{bgm}{Character string giving the name of the atlantis bgm file.}
}
\description{
Transform data from bgm-file to map dataframe.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
bgm <- convert_bgm(dir = d, bgm = "VMPA_setas.bgm")
head(bgm)
}
\seealso{
Other convert functions: \code{\link{convert_factor}},
  \code{\link{convert_path}}, \code{\link{convert_time}}
}

