% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-path.R
\name{convert_path}
\alias{convert_path}
\title{Function to allow for flexible file/filepath usage in the package.}
\usage{
convert_path(dir, file)
}
\arguments{
\item{dir}{Character string giving the directory the file.}

\item{file}{Character string giving the filename.}
}
\value{
Character string of the complete dir/filename string.
}
\description{
Users can either store all their data (modelfiles and outputfiles)
in one directory or use different directories for model data and
output. In case you are using multiple folder please set the argument
\code{dir} to \code{NULL} in all your function calls. In additon
use the complete character string composed of path/file for all
filename-parameters.
}
\examples{
convert_path(dir = file.path("c:", "atlantis-model"), file = "functionalGroups.csv")
convert_path(dir = NULL, file = file.path("c:", "atlantis-model", "functionalGroups.csv"))
}
\seealso{
Other convert functions: \code{\link{convert_bgm}},
  \code{\link{convert_factor}}, \code{\link{convert_time}}
}

