% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{custom_grid}
\alias{custom_grid}
\title{Utility functions used for various plotting routines within atlantistools.}
\usage{
custom_grid(plot, grid_x, grid_y)
}
\arguments{
\item{plot}{ggplot2 plot.}

\item{grid_x}{character vector used in facet_grid in horizontal direction.}

\item{grid_y}{character vector used in facet_grid in vertical direction.}
}
\value{
ggplot2 plot.
}
\description{
Utility functions used for various plotting routines within atlantistools.
}
