% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mort.R
\docType{data}
\name{dir}
\alias{dir}
\title{Plot mortality per group and ageclass.}
\format{An object of class \code{character} of length 1.}
\usage{
dir
}
\arguments{
\item{data}{Dataframe with information about specific mortality The dataframe
should be generated with \code{\link{preproces()}}. Normally this is stored as
\code{peprocess$spec_mort}.}
}
\value{
ggplot2 plot
}
\description{
Plot mortality per group and ageclass.
}
\examples{
plot_mort(pre)
}
\keyword{datasets}

