% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-prm.R
\name{extract_prm}
\alias{extract_prm}
\alias{extract_prm_cohort}
\title{Extract values for Atlantis parameters from the biological parameter file.}
\usage{
extract_prm(dir = getwd(), prm_biol, variables)

extract_prm_cohort(dir = getwd(), prm_biol, variables)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{variables}{Character string giving the flag to search for. This should be
a combination of the parameter name and the group-Code.}
}
\description{
Extract values for Atlantis parameters from the biological parameter file.
}

