% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishbase_data}
\alias{fishbase_data}
\title{fishbase_data}
\format{A data frame with 33104 rows and 12 variables:
\describe{
  \item{\code{SpecCode}}{integer. Species code.}
  \item{\code{Genus}}{character.}
  \item{\code{Species}}{character.}
  \item{\code{SpeciesRefNo}}{integer. Reference number.}
  \item{\code{FBname}}{character. Fishbase name.}
  \item{\code{SubFamily}}{character.}
  \item{\code{FamCode}}{integer. Family Code.}
  \item{\code{GenCode}}{integer. Genetic Code.}
  \item{\code{SubGenCode}}{integer. Sub Genetic Code.}
  \item{\code{Family}}{character.}
  \item{\code{Order}}{character.}
  \item{\code{Class}}{character.}
}}
\source{
\url{http://www.fishbase.org/} rfishbase::fishbase
}
\usage{
fishbase_data
}
\description{
A table of all the the species found in Fihsbase, including taxonomic classification.
}
\keyword{datasets}
