% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-layers.R
\name{flip_layers}
\alias{flip_layers}
\title{flip layers for visualisation}
\usage{
flip_layers(data)
}
\arguments{
\item{data}{dataframe with columns polygon and layer. layer id is based on
atlantis output (0 = layer closest to the sediment)}
}
\value{
dataframe with fliped layerids. 1 = surface.
}
\description{
flip layers for visualisation
}
\examples{
data <- rbind(expand.grid(species = "sp1", polygon = 0, layer = 0:7),
              expand.grid(species = "sp1", polygon = 1, layer = 0:4),
              expand.grid(species = "sp1", polygon = 2, layer = 0:2),
              expand.grid(species = "sp1", polygon = 3, layer = c(0:3, 7)))
data$atoutput <- runif(nrow(data), min = 0, max = 2)
flip_layers(data)
}
