% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-colpal.R
\name{get_colpal}
\alias{get_colpal}
\title{Create discrete color palette used in plots.}
\usage{
get_colpal()
}
\value{
Vector of colors in hexadecimal code.
}
\description{
The colors are derived from http://colorbrewer2.org/ using the palette
12-class Paired. In addition 9 different grey tones were added in
case more than 12 categories are needed (This happens quite often
with feeding data).
}
\seealso{
Other get functions: \code{\link{get_boundary}},
  \code{\link{get_conv_mgnbiot}}, \code{\link{get_groups}}
}
