% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-conv-mgnbiot.R
\name{get_conv_mgnbiot}
\alias{get_conv_mgnbiot}
\title{Extract conversion factor used to transform data from nitrogen in mg to
biomass in tonnes.}
\usage{
get_conv_mgnbiot(dir, prm_biol)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}
}
\value{
Conversion factor as numeric value.
}
\description{
Extract conversion factor used to transform data from nitrogen in mg to
biomass in tonnes.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
get_conv_mgnbiot(dir = d, prm_biol = "VMPA_setas_biol_fishing_Trunk.prm")
}

