% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-growth-fishbase.R
\name{get_growth_fishbase}
\alias{get_growth_fishbase}
\title{Extract growth parameters from www.fishbase.org.}
\usage{
get_growth_fishbase(fish)
}
\arguments{
\item{fish}{Vector of fish species with genus and species information.}
}
\value{
Dataframe with species, country, locality, linf and k.
}
\description{
This function extracts values for Linf, k  and t0 from www.fishbase.org
}
\details{
Before the actual extraction takes place fishbaseh IDs for every species are extracted using \code{\link{get_ids_fishbase}}.
The IDs are needed to generate the urls lateron.
}
\examples{
fish <- c("Gadus morhua", "Merlangius merlangus")
df <- get_growth_fishbase(fish)
head(df)
}
