% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ref-biotoc.R
\name{get_ref_biotic}
\alias{get_ref_biotic}
\title{Extract bibliographic info from www.fishbase.org.}
\usage{
get_ref_biotic(taxon)
}
\arguments{
\item{ref_id}{vector of reference ids.}
}
\value{
Dataframe
}
\description{
Extract bibliographic information for growth parameters (linf, k, t0) from www.fishbase.org
}
\examples{
\dontrun{
taxon <- "Cancer pagurus"
taxon <- "Carcinus maenas"
taxon <- "xxx yyy"
df <- get_ref_biotic(taxon)
}
}
