% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ref-fishbase.R
\name{get_ref_fishbase}
\alias{get_ref_fishbase}
\title{Extract the bibliographic info from www.fishbase.org.}
\usage{
get_ref_fishbase(growth_fishbase, mirror = "se")
}
\arguments{
\item{growth_fishbase}{Dataframe generated with \link{get_growth_fishbase}.}
}
\value{
Dataframe
}
\description{
Extract bibliographic information for growth parameters (linf, k, t0) from www.fishbase.org
}
\examples{
fish <- c("Gadus morhua", "Merlangius merlangus")
growth_fishbase <- get_growth_fishbase(fish)
get_ref_fishbase(growth_fishbase)
growth_fishbase <- get_growth_fishbase("Scyliorhinus canicula")
}
