% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-bps.R
\name{load_bps}
\alias{load_bps}
\title{Extracts the names of the epibenthic biomasspools from the initial conditions file.}
\usage{
load_bps(dir = getwd(), fgs, init)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{init}{Character string giving the filename of netcdf init-file which
shall be read in. Usually "init[...].nc". In case you are using
multiple folder for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}
}
\value{
Character \code{vector} of epibenthic biomass pools.
}
\description{
Use \code{fgs} \code{data.frame} as read in by \code{\link{load_fgs}}
to get the biomass pool information.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
load_bps(dir = d, fgs = "functionalGroups.csv", init = "INIT_VMPA_Jan2015.nc")
}
\seealso{
\code{\link{load_fgs}}

Other load functions: \code{\link{load_box}},
  \code{\link{load_fgs}}, \code{\link{load_nc}}
}

