% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dietcheck.R
\name{load_dietcheck}
\alias{load_dietcheck}
\title{Read in the atlantis dietcheck.txt file and perform some basic data transformations.}
\usage{
load_dietcheck(dir, dietcheck, fgs)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{dietcheck}{Character string of the DietCheck.txt file. Usually
'output[...]DietCheck.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}
}
\value{
A \code{data.frame} in long format with the following coumn names:
  species, time, agecl, and atoutput (i.e., variable).
}
\description{
Read in the atlantis dietcheck.txt file and perform some basic data transformations.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
diet <- load_dietcheck(dir = d, dietcheck = "outputSETASDietCheck.txt", fgs = "functionalGroups.csv")
head(diet, n = 25)
str(diet)
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_fgs}},
  \code{\link{load_nc_physics}}, \code{\link{load_nc}}
}

