% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-init.R
\name{load_init}
\alias{load_init}
\title{This function loads weight at age data (in mgN) from the initial conditions file.}
\usage{
load_init(dir = getwd(), init, vars)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{init}{Character string giving the filename of the initial conditions netcdf file.
Usually "init[...].nc".}

\item{vars}{Vector of character strings giving the variables to extract from the
netcdf file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{select_variable}{Character value spefifying which variable to load.
Only one variable of the options available (i.e., \code{c(
"N", "Nums", "ResN", "StructN", "Eat", "Growth", "Prodn", "Grazing")
}) can be loaded at a time.}

\item{select_groups}{Character vector of funtional groups which shall be read in.
Names have to match the ones used in the ncdf file. Check column "Name" in
"functionalGroups.csv" for clarification.}
}
\value{
A list of dataframes with columns atoutput, polygon and layer (if present).
}
\description{
This function loads weight at age data (in mgN) from the initial conditions file.
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
load_init(dir = dir, init = "init_simple_NorthSea.nc", vars = "cod1_Nums")
load_init(dir = dir, init = "init_simple_NorthSea.nc", vars = c("herring2_ResN", "herring3_ResN"))
load_init(dir = dir, init = "init_simple_NorthSea.nc", vars = "crangon_N")
}
\author{
Alexander Keth
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_fgs}}, \code{\link{load_init_age}},
  \code{\link{load_nc_physics}}, \code{\link{load_nc}},
  \code{\link{load_rec}}, \code{\link{load_spec_mort}},
  \code{\link{load_txt}}
}

