% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-init-age.R
\name{load_init_age}
\alias{load_init_age}
\alias{load_init_nonage}
\alias{load_init_physics}
\alias{load_init_weight}
\title{This function loads weight at age data (in mgN) from the initial conditions file.}
\usage{
load_init_age(dir = getwd(), init, fgs, select_variable,
  select_groups = NULL, bboxes)

load_init_nonage(dir = getwd(), init, fgs, select_variable = "N",
  select_groups = NULL, bboxes, bps)

load_init_physics(dir = getwd(), init, select_variable, bboxes)

load_init_weight(dir = getwd(), init, fgs, bboxes)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{init}{Character string giving the filename of the initial conditions netcdf file.
Usually "init[...].nc".}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{select_variable}{Character value spefifying which variable to load.
For \code{load_init_age} this can be "Nums", "ResN", "StructN",
For \code{load_init_nonage} please select "N" (default)
For \code{load_init_physics} simply pass the names of the physical variables.}

\item{select_groups}{Character vector of funtional groups which shall be read in.
Names have to match the ones used in the ncdf file. Check column "Name" in
"functionalGroups.csv" for clarification.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.
Can be created with \code{get_boundary()}.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file. Can be created with
\code{load_bps()}.}
}
\value{
A dataframes with columns atoutput, polygon, layer (if present), species (if present).
}
\description{
This function loads weight at age data (in mgN) from the initial conditions file.
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
init <- "init_simple_NorthSea.nc"
fgs <- "functionalGroups.csv"
bboxes <- get_boundary(load_box(dir, bgm = "NorthSea.bgm"))
bps = load_bps(dir = dir, fgs = "functionalGroups.csv",
                                init = "init_simple_NorthSea.nc")

load_init_age(dir = dir, init = init, fgs = fgs, bboxes = bboxes,
              select_variable = "ResN",
              select_groups = "cod")

load_init_age(dir = dir, init = init, fgs = fgs, bboxes = bboxes,
              select_variable = "ResN")

load_init_nonage(dir = dir, init = init, fgs = fgs, bboxes = bboxes, bps = bps,
                 select_groups = "crangon")

load_init_nonage(dir = dir, init = init, fgs = fgs, bboxes = bboxes, bps = bps)
}
\author{
Alexander Keth
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_fgs}}, \code{\link{load_init}},
  \code{\link{load_nc_physics}}, \code{\link{load_nc}},
  \code{\link{load_rec}}, \code{\link{load_spec_mort}},
  \code{\link{load_txt}}
}

