% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-mort.R
\name{load_mort}
\alias{load_mort}
\title{Load mortality information from outputMort.txt}
\usage{
load_mort(mortFile, prm_run, fgs)
}
\arguments{
\item{mortFile}{Character string giving the path to the Mort.txt file.
The filename usually contains \code{Mort} and ends in \code{.txt}".}

\item{prm_run}{Character string giving the connection of the run parameterfile.
The filename usually contains \code{run_fishing} and ends in \code{.prm}".}

\item{fgs}{Character string giving the connection to the functional groups file.
The filename usually contains \code{Groups} and does end in \code{.csv}.}
}
\value{
Dataframe with information about sources of mortality (M, F).
}
\description{
Loads Mort.txt file and partitions mortality based on fishing (F) and other mortality (M)
Note: As the Atlantis manual states "This file is currenlty only useful for looking
at \strong{relative M vs F values} for a species, as it does not give accurate mortalities".
Also if a species is set as isImpacted in the \emph{functional_group.csv}, it will have some F value
even if it is not explicity targeted by fishing.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
mortFile <- file.path(d, "outputSETASMort.txt")
prm_run <- file.path(d, "VMPA_setas_run_fishing_F_New.prm")
fgs <- file.path(d, "SETasGroups.csv")

df <- load_mort(mortFile, prm_run, fgs)
head(df)

d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
mortFile <- file.path(d, "outputSETASMort.txt")
prm_run <- file.path(d, "VMPA_setas_run_fishing_F_Trunk.prm")
fgs <- file.path(d, "SETasGroupsDem_NoCep.csv")

df <- load_mort(mortFile, prm_run, fgs)
head(df)
}
\seealso{
Other load functions: 
\code{\link{load_box}()},
\code{\link{load_bps}()},
\code{\link{load_dietcheck}()},
\code{\link{load_fgs}()},
\code{\link{load_init_age}()},
\code{\link{load_init}()},
\code{\link{load_nc_physics}()},
\code{\link{load_nc}()},
\code{\link{load_rec}()},
\code{\link{load_spec_mort}()},
\code{\link{load_txt}()}
}
\concept{load functions}
