% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-nc-physics.R
\name{load_nc_physics}
\alias{load_nc_physics}
\title{Load Atlantis outputfiles (netcdf)}
\usage{
load_nc_physics(dir = getwd(), nc, select_physics, prm_run, aggregate_layers,
  bboxes, warn_zeros = FALSE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc}{Character string giving the filename of netcdf file which
shall be read in. Usually "output[...].nc". Currently the general-
production- and catch.nc files can be loaded in. In case you are using
multiple folder for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{select_physics}{Character vector of physical variables which shall be read in.
Names have to match the ones used in the ncdf file.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.
In addition set dir to 'NULL' in this case.}

\item{aggregate_layers}{Logical indicating if values for layers should be
aggregated (\code{TRUE}) or not (\code{FALSE}).}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{warn_zeros}{Logical indicating if check for actual zeros in the
data shall be printed or not.}
}
\value{
A \code{data.frame} in long format with the following coumn names:
  variable, time, polygon, layer, and atoutput (i.e., variable).
}
\description{
This function loads Atlantis outputfiles (netcdf) and converts them to a dataframe.
}
\details{
This functions converts the ATLANTIS output to a dataframe which can be processed in R.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
test <- load_nc_physics(dir = d, nc = "outputSETAS.nc",
  select_physics = c("salt", "NO3", "volume"),
  prm_run = "VMPA_setas_run_fishing_F_New.prm",
  aggregate_layers = FALSE,
  bboxes = get_boundary(boxinfo = load_box(dir = d, bgm = "VMPA_setas.bgm")))
str(test)
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_fgs}}, \code{\link{load_init_age}},
  \code{\link{load_init}}, \code{\link{load_nc}},
  \code{\link{load_rec}}, \code{\link{load_spec_mort}},
  \code{\link{load_txt}}
}
\keyword{gen}

