% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-rec.R
\name{load_rec}
\alias{load_rec}
\title{Load information for SSB and Recruits from an Atlantis model run.}
\usage{
load_rec(dir = getwd(), yoy, ssb, prm_biol)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{yoy}{Character string of the YOY.txt file. Usually
'output[...]YOY.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{ssb}{Character string of the SSB.txt file. Usually
'output[...]SSB.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}
}
\value{
Dataframe with information about ssb in tonnes and recruits in
thousands.
}
\description{
Load information for SSB and Recruits from an Atlantis model run.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
load_rec(dir = d,
   yoy = "outputSETASYOY.txt",
   ssb = "outputSETASSSB.txt",
   prm_biol = "VMPA_setas_biol_fishing_New.prm")
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_fgs}}, \code{\link{load_init_age}},
  \code{\link{load_init}}, \code{\link{load_nc_physics}},
  \code{\link{load_nc}}, \code{\link{load_spec_mort}},
  \code{\link{load_txt}}
}

