% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-txt.R
\name{load_txt}
\alias{load_txt}
\title{Function to load various txt files from Atlantis simulations}
\usage{
load_txt(dir, file, id_col = "Time")
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{file}{Character string of the file. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as to file. In addition set dir to 'NULL' in this
case.}

\item{id_col}{Character strings giving the names of the columns which are not variables.
Data from all other columns will be gathered with tidyr.}
}
\value{
Dataframe in tidy format!
}
\description{
Function to load various txt files from Atlantis simulations
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
load_txt(dir = d,
   file = "outputSETASSSB.txt",
   id_col = "Time")
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_fgs}}, \code{\link{load_init_weight}},
  \code{\link{load_nc_physics}}, \code{\link{load_nc}},
  \code{\link{load_rec}}, \code{\link{load_spec_mort}}
}

