% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-add-range.R
\name{plot_add_range}
\alias{plot_add_range}
\title{Low level plotting function to add range of observed values to time series plots.}
\usage{
plot_add_range(plot, ex_data)
}
\arguments{
\item{plot}{ggplot2 object.}

\item{ex_data}{Dataframe with observed values for the specific timeseries.}
}
\value{
ggplot2 plot.
}
\description{
This function can be used to add the range of observed data to a timeseries plot
generated with \code{plot_ts()}. The density of the color gives an indication
of the likelihood of the value.
}
\examples{
ex_data <- preprocess_setas$biomass
ex_data$atoutput <- ex_data$atoutput * runif(n = nrow(ex_data), 0, 1)
ex_data$model <- "test"
plot <- plot_ts(preprocess_setas$biomass)
plot <- plot_add_range(plot, ex_data)
plot
}
\seealso{
Other low-level-plot functions: \code{\link{plot_add_run}}
}

