% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-calibrate.R
\name{plot_calibrate}
\alias{plot_calibrate}
\title{Function to plot relative changes over time.}
\usage{
plot_calibrate(data)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Function to plot relative changes over time.
}
\examples{
plot_calibrate(preprocess_setas$structn_age)
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_diet}},
  \code{\link{plot_flux}}, \code{\link{plot_physics}},
  \code{\link{plot_rec}}, \code{\link{plot_species}},
  \code{\link{plot_struct}}, \code{\link{plot_ts}},
  \code{\link{plot_ws}}
}

