% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dietcheck.R
\name{plot_dietcheck}
\alias{plot_dietcheck}
\title{Plot contribution of diet contents for each functional group.}
\usage{
plot_dietcheck(data)
}
\arguments{
\item{data}{Dataframe with information about diets. The dataframe
should be generated with \see{load_dietcheck}.}
}
\value{
ggplot2 object
}
\description{
Plot contribution of diet contents for each functional group.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
plot_diet(dir = d, dietcheck = "outputSETASDietCheck.txt")
}

