% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-flux.R
\name{plot_flux}
\alias{plot_flux}
\title{Function to plot ncdf output of vertical and horizontal fluxes.}
\usage{
plot_flux(data)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Function to plot ncdf output of vertical and horizontal fluxes.
}
\examples{
plot_flux(preprocess_setas$flux)
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}},
  \code{\link{plot_diet_bec_dev}}, \code{\link{plot_diet}},
  \code{\link{plot_line}}, \code{\link{plot_physics}},
  \code{\link{plot_rec}}, \code{\link{plot_species}},
  \code{\link{plot_struct}}, \code{\link{plot_ws}}
}

