% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-line.R
\name{plot_line}
\alias{plot_line}
\title{Function to plot time series of atlantis ncdf output.}
\usage{
plot_line(data, x = "time", y = "atoutput", wrap = "species",
  col = NULL, ncol = 7)
}
\arguments{
\item{data}{Dataframe to be plotted.}

\item{x}{x-variable. Default is \code{'time'}.}

\item{y}{y-variable. Default is \code{'atoutput'}.}

\item{wrap}{Wraping column. Default is \code{'species'}}

\item{col}{Column to use as colour. Default is \code{NULL}.}

\item{ncol}{Number of columns in multipanel plot.}
}
\value{
ggplot2 object
}
\description{
Function to plot time series of atlantis ncdf output.
}
\examples{
plot_line(preprocess_setas$biomass)
plot_line(preprocess_setas$biomass, col = "species")
plot_line(preprocess_setas$biomass_age, col = "agecl")
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_calibrate}},
  \code{\link{plot_diet_bec_dev}}, \code{\link{plot_diet}},
  \code{\link{plot_flux}}, \code{\link{plot_physics}},
  \code{\link{plot_rec}}, \code{\link{plot_species}},
  \code{\link{plot_struct}}, \code{\link{plot_ws}}
}

