% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mort.R
\name{plot_mort}
\alias{plot_mort}
\title{Plot mortality per group and ageclass.}
\usage{
plot_mort(data)
}
\arguments{
\item{data}{Dataframe with information about specific mortality The dataframe
should be generated with \code{\link{preproces()}}. Normally this is stored as
\code{peprocess$spec_mort}.}
}
\value{
ggplot2 plot
}
\description{
Plot mortality per group and ageclass.
}
\examples{
plot_mort(pre)
}

