% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-physics.R
\name{plot_physics}
\alias{plot_physics}
\title{Function to plot ncdf output of physical variables.}
\usage{
plot_physics(data)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Function to plot ncdf output of physical variables.
}
\examples{
plot_physics(preprocess_setas$physics)
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_calibrate}},
  \code{\link{plot_dietcheck}}, \code{\link{plot_diet}},
  \code{\link{plot_flux}}, \code{\link{plot_rec}},
  \code{\link{plot_species}}, \code{\link{plot_struct}},
  \code{\link{plot_ts}}, \code{\link{plot_ws}}
}

