% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rec.R
\name{plot_rec}
\alias{plot_rec}
\title{Plot recruitment.}
\usage{
plot_rec(data, ex_data)
}
\arguments{
\item{data}{Dataframe with information about ssb and recruits.
This is created from atlantis output files YOY.txt and SSB.txt
(Usually output[...]YOY.txt' & 'output[...]SSB.txt') using
\code{\link{load_rec}}.}

\item{ex_data}{Dataframe to compare the atlantis run with.}
}
\value{
ggplot2 object
}
\description{
Plot recruitment.
}
\examples{
data <- preprocess_setas$ssb_rec
ex_data <- data
ex_data$species <- as.character(ex_data$species)
ex_data$rec <- ex_data$rec * runif(n = nrow(ex_data), min = 0.8, max = 1.2)
ex_data$ssb <- ex_data$ssb * runif(n = nrow(ex_data), min = 0.8, max = 1.2)
ex_data$time <- as.character(ex_data$time)
ex_data$model <- "dummy"
plot_rec(data, ex_data)
}

