% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spatial.R
\name{plot_spatial}
\alias{plot_spatial}
\title{Visualise the spatial distribution per species and stanza combination.}
\usage{
plot_spatial(bio_spatial, bgm_as_df, select_species = NULL, timesteps = 2)
}
\arguments{
\item{bio_spatial}{Biomass per group and stanza in tonnes for each timestep,
layer and polygon. This dataframe should be generated with
\code{\link{calculate_biomass_spatial}}. The columns of the dataframe have to
be 'species', 'species_stanza', 'polygon', 'layer', 'time' and 'atoutput'.
Column 'atoutput' is the biomass in tonnes. Please use \code{\link{combine_ages}}
to transform an agebased dataframe to a stanza based dataframe.}

\item{bgm_as_df}{*.bgm file converted to a dataframe. Please use \code{\link{convert_bgm}}
to convert your bgm-file to a dataframe with columns 'lat', 'long', 'inside_lat',
'inside_long' and 'polygon'.}

\item{select_species}{Character vector listing the species to plot. If no species are selected
\code{NULL} (default) all available species are plotted.}

\item{timesteps}{Integer giving the number of timesteps to visualise. The minimum
value is 2 (default). By default the start and end of the simulation is shown. In case
timesteps > 2 equally spaced timesteps - 2 are added.}
}
\value{
grob of 3 ggplot2 plots.
}
\description{
Visualise the spatial distribution per species and stanza combination.
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
nc_gen <- "outputNorthSea.nc"
prm_biol <- "NorthSea_biol_fishing.prm"
prm_run = "NorthSea_run_fishing_F.prm"
bps = load_bps(dir, fgs = "functionalGroups.csv", init = "init_simple_NorthSea.nc")
fgs = "functionalGroups.csv"
bboxes = get_boundary(boxinfo = load_box(dir, bgm = "NorthSea.bgm"))

df_agemat <- prm_to_df(dir = dir, prm_biol = prm_biol, fgs = fgs,
                       group = get_age_acronyms(dir = dir, fgs = fgs),
                       parameter = "age_mat")

df_sp <- calculate_biomass_spatial(dir, nc_gen, prm_biol, prm_run, bps, fgs, bboxes)
bio_spatial <- combine_ages(df_sp, grp_col = "species", agemat = df_agemat)
bgm_as_df <- convert_bgm(dir, bgm = "NorthSea.bgm")

\dontrun{
grobs <- plot_spatial(bio_spatial, bgm_as_df, timesteps = 3)
gridExtra::grid.arrange(grobs[[3]])
}

grobs <- plot_spatial(bio_spatial, bgm_as_df, select_species = "Crangon", timesteps = 3)
gridExtra::grid.arrange(grobs[[1]])
}

