% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spec-mort.R
\name{plot_spec_mort}
\alias{plot_spec_mort}
\title{Plot specific mortality per predator and age!}
\usage{
plot_spec_mort(data, combine_thresh = 0.03)
}
\arguments{
\item{data}{Dataframe with information about dpecific mortality. The dataframe
should be generated with \code{\link{preprocess}}.}

\item{combine_thresh}{Integer indicating minimum amount to the mortality contribution.
Each species item with a lower contribution as this treshold is assigned to the
grpup "Rest". This is necessary for species with a wide large prey spectrum.
Otherwise the screen will be cluttered with colors in the plots. Default is 0.03.}
}
\value{
ggplot2 grobs for multiple pages.
}
\description{
Plot specific mortality per predator and age!
}
\examples{
plots <- plot_spec_mort(data = preprocess_setas$diet_specmort, combine_thresh = 0.03)
gridExtra::grid.arrange(plots[[1]])
}

