% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-struct.R
\name{plot_struct}
\alias{plot_struct}
\title{Function to plot relative contribution of biomass and numbers per cohort.}
\usage{
plot_struct(data)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Function to plot relative contribution of biomass and numbers per cohort.
}
\examples{
plot_struct(preprocess_setas$biomass_age)
}

