% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ts.R, R/plot-ws.R
\name{plot_ts}
\alias{plot_ts}
\title{Function to plot time series of atlantis ncdf output.}
\usage{
plot_ts(data, combine_thresh)

plot_ts(data, combine_thresh)
}
\arguments{
\item{data}{Dataframe to be plotted.}

\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
Function to plot time series of atlantis ncdf output.

Function to plot whole system metrics!
}
\examples{
plot_ts(preprocess_setas$biomass_age)
plot_ws(preprocess_setas$biomass)
}

