% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess-setas.R
\docType{data}
\name{preprocess_setas}
\alias{preprocess_setas}
\title{preprocess-setas}
\format{List with 16 dataframes.
\enumerate{
  \item biomass
  \describe{
    \item{Description}{Biomass of non age-structured groups (< 10 cohorts) calculated as
    nitrogen * volume / dz * conversion_factor for epibenthic groups and
    nitrogen * volume * bio_conv for non epibenthic groups.
    Summed per time and species over polygon and layer. Combined with
    biomass_age Summed per time and species over ages.}
    \item{Columns}{species, time, atoutput}}

  \item biomass_age
  \describe{
    \item{Description}{Biomass as (StructN + ResN) * conversion_factor * Nums for each
    age-structured group (>= 10 cohorts) calculated as summed value
    per time, species and age over polygon and layer.}
    \item{Columns}{species, agecl, time, atoutput}}

  \item diet_dietcheck
  \describe{
    \item{Description}{Indication of feeding interaction derived from output[...]DietCheck.txt file.}
    \item{Columns}{time, pred, habitat, prey, atoutput}}

  \item diet_specmort
  \describe{
    \item{Description}{Indication of feeding interaction derived from output[...]SpecificPredMort.txt file.}
    \item{Columns}{time, prey, agecl, pred, atoutput}}

  \item eat_age
  \describe{
    \item{Description}{"Eat" for each age-structured group calculated as mean value
    per time and species over age and polygon.}
    \item{Columns}{species, time, agecl, atoutput}}

  \item flux
  \describe{
    \item{Description}{Given as raw value for each layer and timestep.}
    \item{Columns}{variable, polygon, layer, time, atoutput}}

  \item grazing
  \describe{
    \item{Description}{"Grazing" for each biomasspool calculated as mean value
    per time and species over polygon.}
    \item{Columns}{species, time, atoutput}}

  \item growth_age
  \describe{
    \item{Description}{"Growth" for each age-structured group calculated as mean value
   per time and species over age and polygon.}
    \item{Columns}{species, time, agecl, atoutput}}

  \item nums
  \describe{
    \item{Description}{"Nums" for each age-structured group calculated as summed value
   per time and species over age, polygon and layer.}
    \item{Columns}{species, time, atoutput}}

  \item nums_age
  \describe{
    \item{Description}{"Nums" for each age-structured group calculated as summed value
    per time, species and polygon over age and layer.}
    \item{Columns}{species, agecl, time, atoutput}}

  \item nums_box
  \describe{
    \item{Description}{Given as mean value over all layers for each physical parameter and time.}
    \item{Columns}{species, polygon, time, atoutput}}

  \item physics
  \describe{
    \item{Description}{}
    \item{Columns}{variable, polygon, time, atoutput}}

  \item resn_age
  \describe{
    \item{Description}{"ResN" for each age-structured group calculated as mean value
   per time, species and age over layer and polygon.}
    \item{Columns}{species, time, agecl, atoutput}}

  \item spec_mort
  \describe{
    \item{Description}{Natural-, predation- and fishing mortality derived from output[...]SpecificMort.txt.}
    \item{Columns}{time, species, agecl, mort, atoutput}}

  \item ssb_rec
  \describe{
    \item{Description}{Spawning stock biomass and recruitment derived from output[...]SSB.txt and output[...]YOY.txt}
    \item{Columns}{species, time, ssb, rec}}

  \item structn_age
  \describe{
    \item{Description}{"StructN" for each age-structured group calculated as mean value
   per time, species and age over layer and polygon.}
    \item{Columns}{species, time, agecl, atoutput}}
    }
\describe{
\item{species}{Name of the functional groups given as character string.
The names match with the column 'LongName' in the functionalGroups.csv file.}
\item{agecl}{Ageclass given as integer from 1 to 10.}
\item{polygon}{Boxid starting from 0 to numboxes - 1.}
\item{layer}{Layerid starting from 0 to numlayers - 1.}
\item{time}{Timestep given as numeric value in years by default. Set \code{as_date} to \code{TRUE} and
provide \code{modelstart} in "yyyy-mm-dd" to convert to date format.}
\item{atoutput}{Obersavtion column storing the actual output value.}
\item{ssb}{Spawning stock biomass in tonnes.}
\item{rec}{Recruits in thousands}
\item{variable}{Used in flux and physics dataframe.}
}}
\usage{
preprocess_setas
}
\description{
preprocess-setas
}
\keyword{datasets}

