% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm-to-df.R
\name{prm_to_df}
\alias{prm_to_df}
\alias{prm_to_df_ages}
\title{Extract parameters from the biological parameter file and transform them to a dataframe.}
\usage{
prm_to_df(prm_biol, fgs, group, parameter)

prm_to_df_ages(prm_biol, fgs, group, parameter)
}
\arguments{
\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm".}

\item{fgs}{Name of the functional groups file which is typically saved as
\code{functionalGroups.csv}.}

\item{group}{Character vector giving the functional Groups to extract.}

\item{parameter}{Character vector giving the parameters to extract.}
}
\value{
Dataframe with columns 'species' and as many columns as parameters.
}
\description{
Extract parameters from the biological parameter file and transform them to a dataframe.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
prm_biol <- file.path(d, "VMPA_setas_biol_fishing_Trunk.prm")
fgs <- file.path(d, "SETasGroupsDem_NoCep.csv")
group <- c("FPS", "FVS")
parameter <- c("mum", "C")

prm_to_df_ages(prm_biol, fgs, group, parameter)
prm_to_df(prm_biol, fgs, group, parameter)
}

