% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nc.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{reexports}
\title{Load Atlantis outputfiles (netcdf)}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc}{Character string giving the filename of netcdf file which
shall be read in. Usually "output[...].nc". Currently the general-
production- and catch.nc files can be loaded in. In case you are using
multiple folder for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{select_groups}{Character vector of funtional groups which shall be read in.
Names have to match the ones used in the ncdf file. Check column "Name" in
"functionalGroups.csv" for clarification.}

\item{select_variable}{Character value spefifying which variable to load.
Only one variable of the options available (i.e., \code{c(
"N", "Nums", "ResN", "StructN", "Eat", "Growth", "Prodn", "Grazing")
}) can be loaded at a time.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}
}
\value{
A \code{data.frame} in long format with the following coumn names:
  species, timestep, polygon, agecl, and atoutput (i.e., variable).
}
\description{
This function loads Atlantis outputfiles (any netcdf file) and
converts them to a \code{data.frame}.
}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}
\author{
Alexander Keth
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_fgs}}
}
\keyword{gen}

