% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_agemat}
\alias{ref_agemat}
\title{agemat.}
\format{\describe{
\item{species}{Name of the functional groups given as character string.
The names match with the column 'LongName' in the functionalGroups.csv file.}
\item{age_mat}{First mature age class.}
}}
\usage{
ref_agemat
}
\description{
agemat.
}
\keyword{datasets}
