% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_bio_cons}
\alias{ref_bio_cons}
\title{Consumed biomass.}
\format{\describe{
\item{pred}{Name of the functional groups given as character string.
The names match with the column 'LongName' in the functionalGroups.csv file.}
\item{agecl}{Ageclass given as integer from 1 to NumCohorts.}
\item{polygon}{Boxid starting from 0 to numboxes - 1.}
\item{time}{Simulation time in years. Modeltimestep was converted to actual
time based on the settings in the 'run.prm' file.}
\item{prey}{Name of the functional groups given as character string.
The names match with the column 'LongName' in the functionalGroups.csv file.}
\item{atoutput}{Obersavtion column storing the actual output value.
Consumed biomass in tonnes.}
}}
\usage{
ref_bio_cons
}
\description{
Consumed biomass.
}
\keyword{datasets}
