% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_dm}
\alias{ref_dm}
\title{Dietcheck.}
\format{\describe{
\item{time}{Simulation time in years. Modeltimestep was converted to actual
time based on the settings in the 'run.prm' file.}
\item{pred}{Name of the functional groups given as character string.
The names match with the column 'LongName' in the functionalGroups.csv file.}
\item{agecl}{Ageclass given as integer from 1 to NumCohorts.}
\item{prey}{Name of the functional groups given as character string.
The names match with the column 'LongName' in the functionalGroups.csv file.}
\item{atoutput}{Obersavtion column storing the actual output value.
Diet contribution in percentage.}
}}
\usage{
ref_dm
}
\description{
Dietcheck.
}
\keyword{datasets}
