% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_physics}
\alias{ref_physics}
\title{Physical variables.}
\format{\describe{
\item{variable}{Name of the physical variable: "salt", "NO3", "NH3", "Temp", "Chl_a" and "Denitrifiction".}
\item{polygon}{Boxid starting from 0 to numboxes - 1.}
\item{layer}{Layerid starting from 0 to numlayers - 1.}
\item{time}{Simulation time in years. Modeltimestep was converted to actual
time based on the settings in the 'run.prm' file.}
\item{atoutput}{Obersavtion column storing the actual output value.
units are salt = PSU; NO3, NH3 = mg N m-3; Temp = degrees Celcius; Chl_a, Denitrifiction = ?}
}}
\usage{
ref_physics
}
\description{
Physical variables.
}
\keyword{datasets}
