% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_to_bibkey.R
\name{ref_to_bibkey}
\alias{ref_to_bibkey}
\title{Convert reference to bib-tex-key.}
\usage{
ref_to_bibkey(ref_df, bib)
}
\arguments{
\item{ref_df}{dataframe with columns author, year, title
in case title is missing match is performed based on author and year only.}

\item{bib}{character string giving the name of the *.bib file.}
}
\value{
Character vector.
}
\description{
Convert reference to bib-tex-key.
}
