% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_vol_dz}
\alias{ref_vol_dz}
\title{Volume and dz.}
\format{\describe{
\item{variable}{Name of the physical variable: "volume" and "dz".}
\item{polygon}{Boxid starting from 0 to numboxes - 1.}
\item{layer}{Layerid starting from 0 to numlayers - 1.}
\item{time}{Simulation time in years. Modeltimestep was converted to actual
time based on the settings in the 'run.prm' file.}
\item{atoutput}{Obersavtion column storing the actual output value.
volume in m^3 dz in m}
}}
\usage{
ref_vol_dz
}
\description{
Volume and dz.
}
\keyword{datasets}
