% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-init.R
\name{sc_init}
\alias{sc_init}
\title{Sanity check initial conditions file}
\usage{
sc_init(dir = getwd(), nc, init, prm_biol, fgs, bboxes)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{init}{Character string giving the filename of the initial conditions netcdf
file Usually "init[...].nc".}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}
}
\value{
Named list with the dataframes as list entry saved as .Rda file.
}
\description{
Sanity check initial conditions file
}

