% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-overlap.R
\name{sc_overlap}
\alias{sc_overlap}
\title{Calculate 3d overlap of predators groups with their prey over time.}
\usage{
sc_overlap(dir = getwd(), nc, prm_biol, bps, fgs, bboxes, out, pred = NULL,
  save_to_disc = FALSE)
}
\description{
Calculate 3d overlap of predators groups with their prey over time.
}

