% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-prm.R
\name{scan_prm}
\alias{extract_prm}
\alias{scan_prm}
\title{Scan character vector for specific flag!}
\usage{
scan_prm(chars, variable)

extract_prm(dir = getwd(), prm_biol, variable)
}
\arguments{
\item{chars}{Vector of character strings}

\item{variable}{Character string giving the flag to search for.}

\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}
}
\description{
Scan character vector for specific flag!
}

