% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-atlantis.R
\name{theme_atlantis}
\alias{theme_atlantis}
\title{Customised theme used in all plots.}
\usage{
theme_atlantis(large = 22, medium = 18, small = 14, scale_font = 1,
  rot_xaxis_text = T, rot_strips_y = T)
}
\arguments{
\item{large}{Integer giving the size of the font for the main parts of the plot.}

\item{medium}{Integer giving the size of the font used in the legend and
facet labels.}

\item{small}{Integer giving the size of the font used in the rest of the plot.}

\item{scale_font}{Numeric used to scale all font sizes. Deafult is 1.}

\item{rot_xaxis_text}{Logical indicating if x-axis text should be
rotated by 45°.}

\item{rot_strips_yLogical}{indicating if facet labels should be
rotated by 90°.}
}
\description{
This function is a customised theme for ggplot2 plots. It's applied
by default to all plots created within \code{atlantistools}.
}
\examples{
nums_agg <- agg_sum(data = ref_nums, groups = c("species", "time"))
ggplot2::ggplot(data = nums_agg, ggplot2::aes(x = time, y = atoutput)) +
   ggplot2::facet_wrap(~species) +
   theme_atlantis()
}

