% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg-mean.R
\name{agg_mean}
\alias{agg_mean}
\alias{agg_sum}
\title{Aggregate data using dplyr functionality.}
\usage{
agg_mean(data, col = "atoutput", groups)

agg_sum(data, col = "atoutput", groups)
}
\arguments{
\item{data}{Dataframe the aggregation is applied to.}

\item{col}{Column of the dataframe the summarise function is applied.
Default is 'atoutput'.}

\item{groups}{Vector of character strings giving the grouping variables.}
}
\value{
grouped datarame with the aggregated data.
}
\description{
This function is a 'wrapper' for the group_by and summarise procedure
used in dplyr to aggregate dataframes.
}
\examples{
agg_ref_nums <- agg_mean(data = ref_nums, groups = c("species", "agecl"))
}

