% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-biomass-spatial.R
\name{calculate_biomass_spatial}
\alias{calculate_biomass_spatial}
\title{Calculate spatiallly explicit biomass (in [t]) for each group and ageclass per timestep.}
\usage{
calculate_biomass_spatial(dir = getwd(), nc_gen, prm_biol, prm_run, bps, fgs,
  bboxes)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc_gen}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}
}
\value{
Dataframe with columns 'species', 'agecl', 'polygon', 'layer', 'time'.
Biomass in [t] is stored in column 'atoutput'.
}
\description{
Calculate spatially explicit biomass time series for each group and ageclass within
our model. Data is read in from 'output[...].nc'. Biomass for age based groups is calculated
as (StructN [mgN] + ResN [mgN]) * Numbers [individuals]. Biomass for non age based groups is
calculated as N [mgN] * volume [m^3] (sediment-dz [m] / volume [m^3] for epibenthic groups).
mgN is converted to t based on the stettings in the biol.prm file. Simulation time steps
are converted to time in years based on output timesteps given in run.prm.
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
nc_gen <- "outputNorthSea.nc"
prm_biol <- "NorthSea_biol_fishing.prm"
prm_run = "NorthSea_run_fishing_F.prm"
bps = load_bps(dir, fgs = "functionalGroups.csv", init = "init_simple_NorthSea.nc")
fgs = "functionalGroups.csv"
bboxes = get_boundary(boxinfo = load_box(dir, bgm = "NorthSea.bgm"))

df <- calculate_biomass_spatial(dir, nc_gen, prm_biol, prm_run, bps, fgs, bboxes)
}

