% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-biomass-spatial.R
\name{calculate_biomass_spatial}
\alias{calculate_biomass_spatial}
\title{Calculate spatiallly explicit biomass (in [t]) for each group and ageclass per timestep.}
\usage{
calculate_biomass_spatial(dir = getwd(), nc_gen, prm_biol, prm_run, bps, fgs,
  bboxes)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc_gen}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}
}
\description{
Calculate spatiallly explicit biomass (in [t]) for each group and ageclass per timestep.
}
\examples{
\dontrun{
dir <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
nc_gen <- "outputSETAS.nc"
prm_biol <- "VMPA_setas_biol_fishing_Trunk.prm"
prm_run = "VMPA_setas_run_fishing_F_Trunk.prm"
bps = load_bps(dir, fgs = "SETasGroupsDem_NoCep.csv", init = "INIT_VMPA_Jan2015.nc")
fgs = "SETasGroupsDem_NoCep.csv"
bboxes = get_boundary(boxinfo = load_box(dir, bgm = "VMPA_setas.bgm"))
df <- calculate_biomass_spatial(dir, nc_gen, prm_biol, prm_run, bps, fgs, bboxes)
}
}

