% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-growth.R
\name{check_growth}
\alias{check_growth}
\title{This function is used to check the individual growth per group over time.}
\usage{
check_growth(data, yearly = FALSE)
}
\arguments{
\item{data}{Dataframe with information about individual age based growth over time.
This should be generated with \code{preprocess()}. You can test either structural or
reserve weight.}

\item{yearly}{Logical specifying if relative change in individual weight shall be calculated
on a yearly basis (\code{TRUE}) ot not (\code{FALSE}). Default is \code{FALSE}.}
}
\value{
Dataframe showing the output of the linear model fit (slope & F-statistic) per
group and age.
}
\description{
This function is used to check the individual growth per group over time.
}
\examples{
check_growth(preprocess$structn_age)
check_growth(preprocess$resn_age)
check_growth(preprocess$resn_age, yearly = TRUE)
}
