% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine-ages.R
\name{combine_ages}
\alias{combine_ages}
\title{Combine ageclasses to juvenile and adult stanza according to age at maturity.}
\usage{
combine_ages(dir = getwd(), data, col, prm_biol)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{data}{Dataframe whose ageclasses shall be combined.}

\item{col}{Character string giving the name of the group column in \code{data}.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}
}
\value{
dataframe with ageclasses combined to stanzas.
}
\description{
Combine ageclasses to juvenile and adult stanza according to age at maturity.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
combine_ages(dir = d,
             data = preprocess_setas$diet_specmort,
             col = "pred",
             prm_biol = "VMPA_setas_biol_fishing_New.prm")
}

