% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-bgm.R
\name{convert_bgm}
\alias{convert_bgm}
\title{Transform data from bgm-file to map dataframe.}
\usage{
convert_bgm(bgm)
}
\arguments{
\item{bgm}{Character string giving the connection to the atlantis bgm file.
The filename ends in \code{.bgm}.}
}
\description{
Transform data from bgm-file to map dataframe.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
bgm <- file.path(d, "VMPA_setas.bgm")

bgm <- convert_bgm(bgm)
head(bgm)
}
\seealso{
Other convert functions: \code{\link{convert_factor}},
  \code{\link{convert_time}}
}
