% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-factor.R
\name{convert_factor}
\alias{convert_factor}
\title{Function to convert any column with information about functional groups
to a factor whose levels use the LongName of the functional groups file.}
\usage{
convert_factor(data_fgs, col)
}
\arguments{
\item{data_fgs}{Dataframe with information about functionalGroups.
Usually loaded with load_fgs().}

\item{col}{Column of the dataframe which is converted to a factor.}
}
\value{
Column of a dataframe in factor format.
}
\description{
This function is used to match the labels of the plots!
}
\seealso{
Other convert functions: \code{\link{convert_bgm}},
  \code{\link{convert_time}}
}
