% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-time.R
\name{convert_time}
\alias{convert_time}
\title{Convert timestep to actual time!}
\usage{
convert_time(dir = getwd(), prm_run, col)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{col}{Numeric vector. Usually a column in a dataframe with information about time.
Either given as timesteps or days.}
}
\value{
Numeric vector with the time in years.
}
\description{
Convert timestep to actual time!
}
\seealso{
Other convert functions: \code{\link{convert_bgm}},
  \code{\link{convert_factor}}, \code{\link{convert_path}}
}

