% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-time.R
\name{convert_time}
\alias{convert_time}
\title{Convert timestep to actual time!}
\usage{
convert_time(dir, prm_run, data, modelstart)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{data}{Dataframe with column having information about the model timestep.}

\item{modelstart}{Character string giving the start of the model run
in the format \code{'yyyy-mm-dd'}.}
}
\value{
Dataframe whose column time is converted from timesteps to
actual time.
}
\description{
Convert timestep to actual time!
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
bps <- load_bps(dir = d, fgs = "functionalGroups.csv", init = "INIT_VMPA_Jan2015.nc")
bboxes <- get_boundary(boxinfo = load_box(dir = d, bgm = "VMPA_setas.bgm"))
test <- load_nc(dir = d, nc = "outputSETAS.nc",
  bps = bps,
  fgs = "functionalGroups.csv",
  select_groups = c("Planktiv_S_Fish", "Cephalopod", "Diatom"),
  select_variable = "Nums",
  bboxes = bboxes,
  check_acronyms = TRUE)
test <- convert_time(dir = d,
  prm_run = "VMPA_setas_run_fishing_F_Trunk.prm",
  data = test,
  modelstart = "1991-01-01")
}

