% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-time.R
\name{convert_time}
\alias{convert_time}
\title{Convert timestep to actual time!}
\usage{
convert_time(dir, prm_run, data, modelstart, stock_state = FALSE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{data}{Dataframe having a column with information about the model timestep.}

\item{modelstart}{Character string giving the start of the model run
in the format \code{'yyyy-mm-dd'}.}

\item{stock_state}{Logical indicating if a stock_state dataframe is changed.}
}
\value{
Dataframe whose column time is converted from timesteps to
actual time.
}
\description{
Convert timestep to actual time!
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
test <- convert_time(dir = d,
   prm_run = "VMPA_setas_run_fishing_F_Trunk.prm",
   data = ref_nums,
   modelstart = "1991-01-01")
head(test)
}

