% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-time.R
\name{convert_time}
\alias{convert_time}
\title{Convert timestep to actual time!}
\usage{
convert_time(prm_run, col)
}
\arguments{
\item{prm_run}{Character string giving the connection of the run parameterfile.
The filename usually contains \code{run_fishing} and ends in \code{.prm}".}

\item{col}{Numeric vector. Usually a column in a dataframe with information about time.
Either given as timesteps or days.}
}
\value{
Numeric vector with the time in years.
}
\description{
Convert timestep to actual time!
}
\seealso{
Other convert functions: \code{\link{convert_bgm}},
  \code{\link{convert_factor}}
}
