% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-groups.R
\name{get_groups}
\alias{get_acronyms}
\alias{get_age_acronyms}
\alias{get_age_groups}
\alias{get_fish_acronyms}
\alias{get_groups}
\alias{get_nonage_acronyms}
\title{Collection of similar functions which get specific
columns from the Atlantis \code{functionalGroups.csv}}
\usage{
get_groups(dir, fgs)

get_age_groups(dir, fgs)

get_acronyms(dir, fgs)

get_age_acronyms(dir, fgs)

get_nonage_acronyms(dir, fgs)

get_fish_acronyms(dir, fgs)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}
}
\value{
Character string.
}
\description{
This collection of functions takes loaded functional
groups, via the \code{\link{load_fgs}} function and creates various
character strings of group names or acronym names.
}
\details{
Currently, the following character strings can be created
- get_groups extracts the column "Name"
- get_age_groups extracts the column "Name". Selects groups with 10 ageclasses.
- get_acronym extracts the column "Code"
- get_age_acronym extracts the column "Code". Selects groups with 10 ageclasses.
- get_nonage_acronym extracts the columne "Code". Only groups with ageclasses different from 10 are selected.
- get_fish_acronyms ectracts the column "Code". Only groups with InvertType equal to "FISH" or "SHARK" are selected.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
get_age_groups(dir = d, fgs = "functionalGroups.csv")
get_nonage_acronyms(dir = d, fgs = "functionalGroups.csv")
}
\seealso{
Other get functions: \code{\link{get_boundary}}
}

