% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-growth-fishbase.R
\name{get_growth_fishbase}
\alias{get_growth_fishbase}
\title{Extract growth parameters from www.fishbase.org.}
\usage{
get_growth_fishbase(fish, exclude_subspecies = T)
}
\arguments{
\item{fish}{Vector of fish species with genus and species information.}
}
\value{
Dataframe with species, country, locality, linf and k.
}
\description{
This function extracts values for Linf and k from www.fishbase.org
}
\details{
Before the actual extraction takes place fishbaseh IDs for every species are extracted using the function "get_ids_fishbase".
The IDs are needed to generate the URLs lateron. At the moment subspecies can only be excluded from the extraction.
}
\examples{
fish <- c("Gadus morhua", "Merlangius merlangus")
extract_linf_k_fishbase(c("Gadus morhua", "Merlangius merlangus"))
}
