% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-ref-biotoc.R
\name{get_ref_biotic}
\alias{get_ref_biotic}
\title{Extract bibliographic info from www.marlin.ac.uk/biotic.}
\usage{
get_ref_biotic(taxon, test = FALSE)
}
\arguments{
\item{taxon}{Character vector of taxon names to search.}

\item{test}{Logical set to \code{TRUE} in case you need to run package development tests. Defaults
to \code{FALSE}.}
}
\value{
Dataframe
}
\description{
Extract bibliographic information for growth, diets, distribution for invertebrate species.
}
\examples{
\dontrun{
taxon <- "Cancer pagurus"
taxon <- "Carcinus maenas"
taxon <- "xxx yyy"
taxon <- "Liocarcinus depurator"
taxon <- "Asterias rubens"
taxon <- "Henricia oculata"
taxon <- "Ensis ensis"
taxon <- "Ampelisca spinipes"
taxon <- "Tubularia indivisa"
taxon <- "Ophelia borealis"

df <- get_ref_biotic(taxon)
}
}
