% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dietcheck.R
\name{load_dietcheck}
\alias{load_dietcheck}
\title{Read in the atlantis dietcheck.txt file and perform some basic data transformations.}
\usage{
load_dietcheck(dir = getwd(), dietcheck)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{dietcheck}{Character string of the DietCheck.txt file. Usually
'output[...]DietCheck.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}
}
\value{
A \code{data.frame} in long format with the following coumn names:
  time, pred, habitat, prey and atoutput (i.e., variable).
}
\description{
Read in the atlantis dietcheck.txt file and perform some basic data transformations.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
diet <- load_dietcheck(dir = d,
    dietcheck = "outputSETASDietCheck.txt")
head(diet, n = 25)
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_fgs}},
  \code{\link{load_nc_physics}}, \code{\link{load_nc}}
}

