% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-dietmatrix.R
\name{load_dietmatrix}
\alias{load_dietmatrix}
\alias{write_diet}
\title{Extract the dietmatrix from the biological parameterfile}
\usage{
load_dietmatrix(dir = getwd(), prm_biol, fgs, transform = TRUE,
  convert_names = FALSE)

write_diet(dir = getwd(), dietmatrix, prm_biol)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{transform}{Boolean indicating if the returned dataframe is displayed in
"long" (\code{transform = TRUE, default}) or "wide" (\code{transform = FALSE})
format. You should use the "wide" format in case you aim to change your
diet matrix entries.}

\item{convert_names}{Logical indicating if group codes are transformed to LongNames (\code{TRUE})
or not (default = \code{FALSE}).
'avail'. The dataframe should be generated with \code{load_dietmatrix()}.}

\item{dietmatrix}{Dataframe in 'long' format containing information about availabilities
with columns 'pred', 'prey', 'pred_stanza', 'prey_stanza', 'code', 'prey_id' and}
}
\value{
dataframe of the availability matrix in long format with columns
pred, pred_stanza (1 = juvenile, 2 = adult), prey_stanza, prey, avail, code.
}
\description{
Extracts the diet matrix as long dataframe from the biological paremeter file
of any ATLANTIS simulation.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
dm <- load_dietmatrix(dir = d,
                      prm_biol = "VMPA_setas_biol_fishing_New.prm",
                      fgs = "SETasGroups.csv")
head(dm, n = 10)
}

