% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-fgs.R
\name{load_fgs}
\alias{load_fgs}
\title{Load the functional group file}
\usage{
load_fgs(dir = getwd(), fgs)
}
\arguments{
\item{dir}{Path of the Atlantis model folder.}

\item{fgs}{Name of the functional groups file which is typically saved as
\code{functionalGroups.csv}.}
}
\value{
A \code{data.frame} of functional group information.
}
\description{
Read in the functional group file as dataframe.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
file <- "functionalGroups.csv"
fgs <- load_fgs(d, file)
head(fgs)
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_nc_physics}}, \code{\link{load_nc}}
}

