% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-nc.R
\name{load_nc}
\alias{load_nc}
\title{Load Atlantis outputfiles (netcdf)}
\usage{
load_nc(dir = getwd(), nc, fgs, bps, select_groups, select_variable, prm_run,
  bboxes = c(0), check_acronyms = TRUE, warn_zeros = FALSE,
  report = TRUE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc}{Character string giving the filename of netcdf file which
shall be read in. Usually "output[...].nc". Currently the general-
production- and catch.nc files can be loaded in. In case you are using
multiple folder for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{select_groups}{Character vector of funtional groups which shall be read in.
Names have to match the ones used in the ncdf file. Check column "Name" in
"functionalGroups.csv" for clarification.}

\item{select_variable}{Character value spefifying which variable to load.
Only one variable of the options available (i.e., \code{c(
"N", "Nums", "ResN", "StructN", "Eat", "Growth", "Prodn", "Grazing")
}) can be loaded at a time.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{check_acronyms}{Logical testing if functional-groups in
select_groups are inactive in the current model run. The will be omitted
in the output.}

\item{warn_zeros}{Logical indicating if check for actual zeros in the
data shall be printed or not.}

\item{report}{Logical indicating if progress bars shall be printed (\code{TRUE}) or
not (\code{FALSE}).}
}
\value{
A \code{data.frame} in long format with the following column names:
  species, timestep, polygon, agecl, and atoutput (i.e., variable).
}
\description{
This function loads Atlantis outputfiles (any netcdf file) and
converts them to a \code{data.frame}.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
nc <- "outputSETAS.nc"
bps <- load_bps(dir = d, fgs = "SETasGroupsDem_NoCep.csv", init = "INIT_VMPA_Jan2015.nc")
fgs <- "SETasGroupsDem_NoCep.csv"
bboxes <- get_boundary(boxinfo = load_box(dir = d, bgm = "VMPA_setas.bgm"))
prm_run <- "VMPA_setas_run_fishing_F_Trunk.prm"

test <- load_nc(dir = d, nc = nc, bps = bps, fgs = fgs, prm_run = prm_run, bboxes = bboxes,
  select_groups = c("Planktiv_S_Fish", "Cephalopod", "Diatom"),
  select_variable = "ResN")

test <- load_nc(dir = d, nc = nc, bps = bps, fgs = fgs, prm_run = prm_run, bboxes = bboxes,
  select_groups = c("Planktiv_S_Fish", "Cephalopod", "Diatom"),
  select_variable = "Nums")
}
\author{
Alexander Keth
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_fgs}}, \code{\link{load_init_age}},
  \code{\link{load_init}}, \code{\link{load_nc_physics}},
  \code{\link{load_rec}}, \code{\link{load_spec_mort}},
  \code{\link{load_txt}}
}
\keyword{gen}

