% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-spec-mort.R
\name{load_spec_mort}
\alias{load_spec_mort}
\title{Load mortality information from outputSpecificPredMort.txt}
\usage{
load_spec_mort(dir = getwd(), specmort, prm_run, convert_names = FALSE,
  version_flag = 1)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{specmort}{Character string of the outputSpecificPredMort.txt file. Usually
'output[...]SpecificPredMort.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.
In addition set dir to 'NULL' in this case.}

\item{convert_names}{Logical indicating if group codes are transformed to LongNames (\code{TRUE})
or not (default = \code{FALSE}).}

\item{version_flag}{The version of atlantis that created the output files.
1 for bec_dev, 2 for trunk.}
}
\value{
Dataframe with information about ssb in tonnes and recruits in
thousands.
}
\description{
Load mortality information from outputSpecificPredMort.txt
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
df <- load_spec_mort(dir = d,
                     specmort = "outputSETASSpecificPredMort.txt",
                     prm_run = "VMPA_setas_run_fishing_F_New.prm")
head(df)

d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
df <- load_spec_mort(dir = d,
                     specmort = "outputSETASSpecificPredMort.txt",
                     prm_run = "VMPA_setas_run_fishing_F_Trunk.prm",
                     version_flag = 2)
head(df)
}
\seealso{
Other load functions: \code{\link{load_box}},
  \code{\link{load_bps}}, \code{\link{load_dietcheck}},
  \code{\link{load_fgs}}, \code{\link{load_init_age}},
  \code{\link{load_init}}, \code{\link{load_nc_physics}},
  \code{\link{load_nc}}, \code{\link{load_rec}},
  \code{\link{load_txt}}
}

