% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-spec-mort.R
\name{load_spec_mort}
\alias{load_spec_mort}
\title{Load mortality information from outputSpecificPredMort.txt}
\usage{
load_spec_mort(dir = getwd(), specmort)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{specmort}{Character string of the outputSpecificPredMort.txt file. Usually
'output[...]SpecificPredMort.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}
}
\value{
Dataframe with information about ssb in tonnes and recruits in
thousands.
}
\description{
Load mortality information from outputSpecificPredMort.txt
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
load_spec_mort(dir = d,
   specmort = "outputSETASSpecificPredMort.txt")
}

