% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-add-run.R
\name{plot_add_run}
\alias{plot_add_run}
\title{Low level plotting function to apply color coding to plots based on run variable!}
\usage{
plot_add_run(plot)
}
\arguments{
\item{plot}{ggplot2 object.}
}
\value{
ggplot2 plot.
}
\description{
This function can be used to use plotting routines without color specification
e.g. `plot_ts()`, `plot_calibrate()` or `plot_physics()` for model comparisons.
Please note that `plot_ts()` and `plot_calibrate()` only work for non age based
data. The colorcoding is based on the column run so you need to apply combine runs
to your data first!
}
\examples{
dummy <- preprocess_setas
# Apply
dummy$biomass$atoutput <- dummy$biomass$atoutput * 1.5
}

