% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bench.R
\name{plot_bench}
\alias{plot_bench}
\title{Compare Atlantis scenario with other models/data sources.}
\usage{
plot_bench(data, dir, compare)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Compare Atlantis scenario with other models/data sources.
}
\examples{
plot_bench(preprocess_setas$structn_age)
}

