% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bench.R
\name{plot_bench}
\alias{plot_bench}
\title{Compare Atlantis scenario with other models/data sources.}
\usage{
plot_bench(data, ex_data)
}
\arguments{
\item{data}{Dataframe to be plotted.}

\item{ex_data}{Dataframe to compare the atlantis run with.}
}
\value{
ggplot2 object
}
\description{
Compare Atlantis scenario with other models/data sources.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
ex_data <- read.csv(file.path(d, "setas-bench.csv"), stringsAsFactors = FALSE)
plot_bench(data = preprocess_setas$biomass, ex_data)
}
\seealso{
Other plot functions: \code{\link{plot_boxes}},
  \code{\link{plot_calibrate}}, \code{\link{plot_diet}},
  \code{\link{plot_flux}}, \code{\link{plot_physics}},
  \code{\link{plot_rec}}, \code{\link{plot_species}},
  \code{\link{plot_struct}}, \code{\link{plot_ts}},
  \code{\link{plot_ws}}
}

