% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-calibrate.R
\name{plot_calibrate}
\alias{plot_calibrate}
\title{Function to plot relative changes over time.}
\usage{
plot_calibrate(data)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Function to plot relative changes over time.
}
\examples{
plot_calibrate(preprocess_setas$structn_age)
}

