% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-consumed-biomass.R
\name{plot_consumed_biomass}
\alias{plot_consumed_biomass}
\title{Circle diagramm to visualise the consumed biomass for the whole system.}
\usage{
plot_consumed_biomass(df, select_time, show)
}
\description{
Circle diagramm to visualise the consumed biomass for the whole system.
}

