% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diet.R
\name{plot_diet}
\alias{plot_diet}
\title{Plot contribution of diet contents for each functional group.}
\usage{
plot_diet(preddata, preydata, species = NULL, wrap_col, combine_thresh = 7)
}
\arguments{
\item{preddata}{Diet proportion of prey species in the stomach of a specific
predator/ageclass combination per time. This dataframe
should be generated with \code{\link{load_dietcheck}} or read in from a preprocessd
*.rda generated with \code{\link{preprocess}}.}

\item{preydata}{Proportion of consumed biomass of a preygroup by different predator/ageclass
combinations. This dataframe should be generated with \code{\link{biomass_flow}} with
\code{plot_diet} set to \code{TRUE}.}

\item{species}{Character string giving the acronyms of the species you aim to plot. Default is
\code{NULL} resulting in all available species being ploted.}

\item{wrap_col}{Character specifying the column of the dataframe to be used as multipanel plot.
Currently only "agecl" is supported.}

\item{combine_thresh}{Number of different categories to plot. Lets say predator X has eaten
20 different prey items. If you only want to show the 3 most important prey items set
\code{combine_thresh} to 3. As rule of thumb values < 10 are useful otherwise to many
colors are used in the plots.}
}
\value{
List of ggplot2 objects.
}
\description{
Visualise diet proportions form predator and prey perspective. The upper panel
plot shows the predator perspective while the lower panel plot shows the prey perspective
for a given group. Please note that this function only works properly with models
based on the trunk code. Bec_dev models should use \code{plot_diet_bec_dev} to get an indication
of the feeding interactions.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_calibrate}},
  \code{\link{plot_diet_bec_dev}}, \code{\link{plot_flux}},
  \code{\link{plot_physics}}, \code{\link{plot_rec}},
  \code{\link{plot_species}}, \code{\link{plot_struct}},
  \code{\link{plot_ts}}, \code{\link{plot_ws}}
}

