% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diet.R
\name{plot_diet}
\alias{plot_diet}
\title{Plot contribution of diet contents for each functional group.}
\usage{
plot_diet(data, species = NULL, wrap_col, combine_thresh = 15)
}
\arguments{
\item{data}{Dataframe with information about diets. The dataframe
should be generated with \code{\link{load_dietcheck}}.}

\item{species}{Character string giving the acronyms of the species you aim to plot. Default is
\code{NULL} resulting in all available species being ploted.}

\item{wrap_col}{Character specifying the column of the dataframe to be used as multipanel plot.
In case you aim to plot DietCheck.txt data use "habitat".
In case you aim to plot SpecMort.txt data use either "agecl" or "stanza".}

\item{combine_thresh}{Integer indicating minimum amount to the stomach contribution.
Each prey item with a lower contribution as this treshold is assigned to the
grpup "Rest". This is necessary for species with a wide variety of food items
in their diet. Otherwise the screen will be cluttered with colors in the
dietplots. Default is 0.03.}
}
\value{
List of ggplot2 objects.
}
\description{
Plot contribution of diet contents for each functional group.
}
\examples{
# Plot DietCheck.txt
plots <- plot_diet(preprocess_setas$diet_dietcheck, wrap_col = "habitat")
gridExtra::grid.arrange(plots[[1]])

# Plot SpecMort.txt per stanza First we need to transform the ageclasses to stanzas.
\dontrun{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
diet_stanza <- combine_ages(dir = d,
                            data = preprocess_setas$diet_specmort,
                            col = "pred",
                            prm_biol = "VMPA_setas_biol_fishing_New.prm")
plots <- plot_diet(diet_stanza, wrap_col = "stanza")
gridExtra::grid.arrange(plots[[1]])
}

# Plot SpecMort.txt per ageclass.
plots <- plot_diet(preprocess_setas$diet_specmort, wrap_col = "agecl")
gridExtra::grid.arrange(plots[[1]])

# Only plot specific species
plots <- plot_diet(preprocess_setas$diet_specmort, species = "CEP", wrap_col = "agecl")
gridExtra::grid.arrange(plots[[1]])
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_calibrate}},
  \code{\link{plot_dietcheck}}, \code{\link{plot_flux}},
  \code{\link{plot_physics}}, \code{\link{plot_rec}},
  \code{\link{plot_species}}, \code{\link{plot_struct}},
  \code{\link{plot_ts}}, \code{\link{plot_ws}}
}

