% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dietcheck.R
\name{plot_diet}
\alias{plot_diet}
\title{Plot contribution of diet contents for each functional group.}
\usage{
plot_diet(dir = getwd(), dietcheck)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis output folder.}

\item{dietcheck}{Character string of the DietCheck.txt file. Usually
'output[...]DietCheck.txt'.}
}
\value{
ggplot2 object
}
\description{
Plot contribution of diet contents for each functional group.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
plot_diet(dir = d, dietcheck = "outputSETASDietCheck.txt")
}

