% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dietcheck.R
\name{plot_dietcheck}
\alias{plot_dietcheck}
\title{Plot contribution of diet contents for each functional group.}
\usage{
plot_dietcheck(data, combine_thresh = 0.03)
}
\arguments{
\item{data}{Dataframe with information about diets. The dataframe
should be generated with \code{\link{load_dietcheck}}.}

\item{combine_thresh}{Integer indicating minimum amount to the stomach contribution.
Each prey item with a lower contribution as this treshold is assigned to the
grpup "Rest". This is necessary for species with a wide variety of food items
in their diet. Otherwise the screen will be cluttered with colors in the
dietplots. Default is 0.03.}
}
\value{
List of ggplot2 objects.
}
\description{
Plot contribution of diet contents for each functional group.
}
\examples{
plots <- plot_dietcheck(data = preprocess_setas$diet_dietcheck, combine_thresh = 0.03)
gridExtra::grid.arrange(plots[[1]])
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_calibrate}},
  \code{\link{plot_diet}}, \code{\link{plot_flux}},
  \code{\link{plot_physics}}, \code{\link{plot_rec}},
  \code{\link{plot_struct}}, \code{\link{plot_ts}},
  \code{\link{plot_ws}}
}

