% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rec.R
\name{plot_rec}
\alias{plot_rec}
\title{Plot recruitment.}
\usage{
plot_rec(data, ex_data)
}
\arguments{
\item{data}{Dataframe with information about ssb and recruits.
This is created from atlantis output files YOY.txt and SSB.txt
(Usually output[...]YOY.txt' & 'output[...]SSB.txt') using
\code{\link{load_rec}}.}

\item{ex_data}{Dataframe to compare the atlantis run with.}
}
\value{
ggplot2 object
}
\description{
Plot recruitment.
}
\examples{
d <- system.file("extdata", "setas-model-new-trunk", package = "atlantistools")
data <- load_rec(dir = d,
   yoy = "outputSETASYOY.txt",
   ssb = "outputSETASSSB.txt",
   fgs = "functionalGroups.csv",
   prm_biol = "VMPA_setas_biol_fishing_Trunk.prm",
   prm_run = "VMPA_setas_run_fishing_F_Trunk.prm",
   modelstart = "1991-01-01")

ex_data <- data
ex_data$species <- as.character(ex_data$species)
ex_data$rec <- ex_data$rec * runif(n = nrow(ex_data), min = 0.8, max = 1.2)
ex_data$ssb <- ex_data$ssb * runif(n = nrow(ex_data), min = 0.8, max = 1.2)
ex_data$model <- "test_model"
plot_rec(data, ex_data)
}

