% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rec.R
\name{plot_rec}
\alias{plot_rec}
\title{Plot recruitment.}
\usage{
plot_rec(data, ex_data)
}
\arguments{
\item{data}{Dataframe with information about ssb and recruits.
This is created from atlantis output files YOY.txt and SSB.txt
(Usually output[...]YOY.txt' & 'output[...]SSB.txt') using
\code{\link{load_rec}}.}

\item{ex_data}{Dataframe to compare the atlantis run with.}
}
\value{
ggplot2 object
}
\description{
Plot recruitment.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
ex_data <- read.csv(file.path(d, "setas-ssb-rec.csv"), stringsAsFactors = FALSE)
plot_rec(preprocess_setas$ssb_rec, ex_data)
}

