% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spatial-overlap.R
\name{plot_spatial_overlap}
\alias{plot_spatial_overlap}
\title{Plot spatial overlap.}
\usage{
plot_spatial_overlap(df_list)
}
\arguments{
\item{df_list}{List of dataframes generated with \code{\link{calculate_spatial_overlap}}}
}
\value{
ggplot2 plot.
}
\description{
Plot spatial overlap.
}
\examples{
sp_overlap <- calculate_spatial_overlap(ref_bio_sp, ref_dietmatrix, ref_agemat)
plot_spatial_overlap(sp_overlap)
}
