% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-species.R
\name{plot_species}
\alias{plot_species}
\title{Create species specific overvie plot.}
\usage{
plot_species(data_pre, species)
}
\arguments{
\item{data_pre}{List of preprocessed Atlantis simulation. The list of dataframes
should be created with \code{preprocess()}.}

\item{species}{Character srtring giving the name of the species to plot. Only age
based species are supported.}
}
\value{
ggplot2 object of class grob
}
\description{
This plotting routine is based on Raphael's (Ifremer) plotting routune
used during model calibration. Currently 4 plots are created by default:
- StructN over time per age
- ResN over time per age
- Condition over time per age
- Numbers over time per age
}
\examples{
plot <- plot_species(preprocess_setas, species = "Shallow piscivorous fish")
# Use grid.arrange to draw the plot on the current device
gridExtra::grid.arrange(plot)
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_calibrate}},
  \code{\link{plot_dietcheck}}, \code{\link{plot_diet}},
  \code{\link{plot_flux}}, \code{\link{plot_physics}},
  \code{\link{plot_rec}}, \code{\link{plot_struct}},
  \code{\link{plot_ts}}, \code{\link{plot_ws}}
}

