% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Function to plot time series of atlantis ncdf output.}
\usage{
plot_ts(data, x = "time", y = "atoutput", wrap = "species", ncol = 7)
}
\arguments{
\item{data}{Dataframe to be plotted.}

\item{x}{x-variable. Default is \code{'time'}.}

\item{y}{y-variable. Default is \code{'atoutput'}.}

\item{wrap}{Wraping column. Default is \code{'species'}}

\item{ncol}{Number of columns in multipanel plot.}
}
\value{
ggplot2 object
}
\description{
Function to plot time series of atlantis ncdf output.
}
\examples{
plot_ts(preprocess_setas$biomass_age)
}
\seealso{
Other plot functions: \code{\link{plot_boxes}},
  \code{\link{plot_calibrate}}, \code{\link{plot_diet}},
  \code{\link{plot_flux}}, \code{\link{plot_physics}},
  \code{\link{plot_rec}}, \code{\link{plot_species}},
  \code{\link{plot_struct}}, \code{\link{plot_ws}}
}

