% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Function to plot time series of atlantis ncdf output.}
\usage{
plot_ts(data)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Function to plot time series of atlantis ncdf output.
}
\examples{
plot_ts(preprocess_setas$biomass)
}

