% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ws.R
\name{plot_ws}
\alias{plot_ws}
\title{Function to plot whole system metrics!}
\usage{
plot_ws(data, combine_thresh = 0.03)
}
\arguments{
\item{data}{Dataframe to be plotted.}
}
\value{
ggplot2 object
}
\description{
Function to plot whole system metrics!
}
\examples{
plot_ws(preprocess_setas$biomass)
}

