% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-ws.R
\name{plot_ws}
\alias{plot_ws}
\title{Function to plot whole system metrics!}
\usage{
plot_ws(data, combine_thresh = 15)
}
\arguments{
\item{data}{Dataframe to be plotted.}

\item{combine_thresh}{Integer indicating the number of groups to display. Default is \code{15}.}
}
\value{
ggplot2 object
}
\description{
Function to plot whole system metrics!
}
\examples{
plot_ws(preprocess_setas$biomass)
}
\seealso{
Other plot functions: \code{\link{plot_bench}},
  \code{\link{plot_boxes}}, \code{\link{plot_calibrate}},
  \code{\link{plot_diet_bec_dev}}, \code{\link{plot_diet}},
  \code{\link{plot_flux}}, \code{\link{plot_line}},
  \code{\link{plot_physics}}, \code{\link{plot_rec}},
  \code{\link{plot_species}}, \code{\link{plot_struct}}
}

