% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess Atlantis output (netcdf)}
\usage{
preprocess(dir = getwd(), nc_gen, nc_prod, dietcheck, yoy, ssb, prm_biol,
  prm_run, bps, fgs, select_groups, bboxes, check_acronyms, modelstart, out,
  report = TRUE, save_to_disc = FALSE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc_gen}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{nc_prod}{Character string giving the filename of the productivity netcdf
file. Usually "output[...]PROD.nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_prod. In addition set dir to 'NULL' in this
case.}

\item{dietcheck}{Character string of the DietCheck.txt file. Usually
'output[...]DietCheck.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as dietcheck. In addition set dir to 'NULL' in this
case.}

\item{yoy}{Character string of the YOY.txt file. Usually
'output[...]YOY.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as yoy. In addition set dir to 'NULL' in this
case.}

\item{ssb}{Character string of the SSB.txt file. Usually
'output[...]SSB.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as ssb. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{prm_run}{Character string giving the filename of the run
parameterfile. Usually "[...]run_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{select_groups}{Character vector of funtional groups which shall be read in.
Names have to match the ones used in the ncdf file. Check column "Name" in
"functionalGroups.csv" for clarification.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{check_acronyms}{Logical testing if functional-groups in
select_groups are inactive in the current model run. The will be omitted
in the output.}

\item{modelstart}{Character string giving the start of the model run
in the format \code{'yyyy-mm-dd'}.}

\item{out}{Character string giving the filename of the *.Rda output
file. In case you are using different folders for your model data
and output files please pass a complete folder/filename string and
set dir to 'NULL'.}

\item{report}{Logical indicating if additional information shall be printed
during function evaluation.}

\item{save_to_disc}{Logical indicating if the resulting list shall be stored
on the hard-disc (\code{TRUE}) or not (\code{FALSE}).}

\item{specmort}{Character string of the SpecMort.txt file. Usually
'output[...]SpecificPredMort.txt'. In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as specmort. In addition set dir to 'NULL' in this
case.}
}
\value{
Named list with the dataframes as list entry saved as .Rda file.
}
\description{
This function loads data from Atlantis output (netcdf) files, performs various
calculations and saves the intermediate output as .Rda file. Information
about individual weight is extracted from the general netcdf output file via
"N", "ResN", "StructN" in additon information about numbers is extracted via
"Nums". Using the productivity outputfile information about feeding is extracted
("Eat", "Grazing"). The raw-data is aggregated in various levels of
complexity. E.g. spatially, age-based, vertically.
Please see the details section for further information about the specific
calculations.
}
\details{
This functions performs various calculations (details below) and saves the result as a
named list of dataframes.In addition results are saved to HDD as *.Rda file.

1. structn_age
   "StructN" for each age-structured group calculated as mean value
   per time, species and age over layer and polygon.
2. resn_age
   "ResN" for each age-structured group calculated as mean value
   per time, species and age over layer and polygon.
3. eat_age
   "Eat" for each age-structured group calculated as mean value
   per time, species and age over polygon.
4. growth_age
   "Growth" for each age-structured group calculated as mean value
   per time, species and age over polygon.
7. eat
   "Eat" for each age-structured group calculated as mean value
   per time and species over age and polygon.
8. growth
   "Growth" for each age-structured group calculated as mean value
   per time and species over age and polygon.
9. numbers
   "Nums" for each age-structured group calculated as summed value
   per time and species over age, polygon and layer.
10. numbers_age
    "Nums" for each age-structured group calculated as summed value
    per time, species and age over polygon and layer.
11. numbers_box
    "Nums" for each age-structured group calculated as summed value
    per time, species and polygon over age and layer.
12. physics
    Given as mean value over all layers for each physical parameter and time.
13. flux_layer
    Given as raw value for each layer and timestep.
14. grazing
    "Grazing" for each biomasspool calculated as mean value
    per time and species over polygon.
15. biomass_age
    Biomass as (StructN + ResN) * conversion_factor * Nums for each
    age-structured group (>= 10 cohorts) calculated as summed value
    per time, species and age over polygon and layer.
16. biomass
    Biomass of non age-structured groups (< 10 cohorts) calculated as
    nitrogen * volume / dz * conversion_factor for epibenthic groups and
    nitrogen * volume * bio_conv for non epibenthic groups.
    Summed per time and species over polygon and layer. Combined with
    biomass_age Summed per time and species over ages.
}
\examples{
d <- system.file("extdata", "setas-model-new-becdev", package = "atlantistools")
test <- preprocess(dir = d,
   nc_gen = "outputSETAS.nc",
   nc_prod = "outputSETASPROD.nc",
   dietcheck = "outputSETASDietCheck.txt",
   yoy = "outputSETASYOY.txt",
   ssb = "outputSETASSSB.txt",
   prm_biol = "VMPA_setas_biol_fishing_New.prm",
   prm_run = "VMPA_setas_run_fishing_F_New.prm",
   bps = load_bps(dir = d, fgs = "SETasGroups.csv", init = "init_vmpa_setas_25032013.nc"),
   fgs = "SETasGroups.csv",
   select_groups = c("Planktiv_S_Fish", "Pisciv_S_Fish", "Cephalopod", "Diatom", "Zoo"),
   bboxes = get_boundary(boxinfo = load_box(dir = d, bgm = "VMPA_setas.bgm")),
   check_acronyms = TRUE,
   modelstart = "1991-01-01",
   out = "preprocess.Rda",
   save_to_disc = FALSE)
}
\keyword{gen}

