% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess.R
\docType{data}
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\format{List with 17 dataframes.
\enumerate{
  \item biomass
  \item biomass_age
  \item biomass_consumed
  \item biomass_spatial_stanza
  \item diet
  \item eat_age
  \item flux
  \item grazing
  \item growth_age
  \item growth_rel_init
  \item nums
  \item nums_age
  \item nums_box
  \item physics
  \item resn_age
  \item sink
  \item spatial_overlap
  \item ssb_rec
  \item structn_age
  \item vol
    }}
\usage{
preprocess
}
\description{
See "data-vignette-model-preprocess.R" and "model-preprocess.Rmd" for further information.
}
\keyword{datasets}
