% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocess-setas.R
\docType{data}
\name{preprocess_setas}
\alias{preprocess_setas}
\title{preprocess-setas}
\format{List with 15 dataframes.}
\source{
See \code{data-raw/data-preprocess.R} for further information.
}
\usage{
preprocess_setas
}
\description{
preprocess-setas
}
\keyword{datasets}

