% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm-to-df.R
\name{prm_to_df}
\alias{prm_to_df}
\alias{prm_to_df_ages}
\title{Extract parameters from the biological parameter file and transform them to a dataframe.}
\usage{
prm_to_df(dir = getwd(), prm_biol, fgs, group, parameter)

prm_to_df_ages(dir = getwd(), prm_biol, fgs, group, parameter)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{group}{Character vector giving the functional Groups to extract.}

\item{parameter}{Character vector giving the parameters to extract.}
}
\value{
Dataframe with columns 'species' and as many columns as parameters.
}
\description{
Extract parameters from the biological parameter file and transform them to a dataframe.
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
prm_biol <- "NorthSea_biol_fishing.prm"
fgs <- "functionalGroups.csv"
group <- c("COD", "HER")

prm_to_df_ages(dir, prm_biol, fgs, group, parameter = c("mum", "C"))
prm_to_df(dir, prm_biol, fgs, group, parameter = c("age_mat", "KWRR"))
}

