% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-init.R
\name{sc_init}
\alias{plot_sc_init}
\alias{sc_init}
\title{Sanity check initial conditions file}
\usage{
sc_init(dir = getwd(), init, prm_biol, fgs, bboxes, out, pred = NULL,
  set_avail = NULL, save_to_disc = FALSE)

plot_sc_init(df, mult_mum, mult_c, pred = NULL)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{init}{Character string giving the filename of the initial conditions netcdf
file Usually "init[...].nc".}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{out}{Character string giving the filename of the *.Rda output
file. In case you are using different folders for your model data
and output files please add the output folder.}

\item{pred}{Vector of predator acronyms to check. If \code{NULL} (default) all age based
predators are selected.}

\item{set_avail}{Numeric value. All present availabilities can be set to a spefiic value.
Default value is \code{NULL} which results in no changes to the present availability matrix.}

\item{save_to_disc}{Logical indicating if the resulting list shall be stored
on the hard-disc (\code{TRUE}) or not (\code{FALSE}).}

\item{df}{Dataframe to pass to \code{plot_sc_init()}. df should be generated with
sc_init or read in from *.rda (also generated with sc_init()).}

\item{mult_mum}{Numeric vector of multiplication factors applied to the initial
mum values.}

\item{mult_c}{Numeric vector of multiplication factors applied to the initial
C values.}
}
\value{
Dataframe/ Plot.
}
\description{
Sanity check initial conditions file
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
fgs <- "functionalGroups.csv"
init <- "init_simple_NorthSea.nc"
prm_biol <- "NorthSea_biol_fishing.prm"
bboxes <- get_boundary(load_box(dir = dir, bgm = "NorthSea.bgm"))
mult_mum <- seq(0.5, 10, by = 1)
mult_c <- seq(0.5, 10, by = 1)
no_avail <- FALSE
save_to_disc <- FALSE
data1 <- sc_init(dir, init, prm_biol, fgs, bboxes, save_to_disc = FALSE)
plot_sc_init(df = data1, mult_mum, mult_c)
plot_sc_init(df = data1, mult_mum, mult_c, pred = "Cod")

data2 <- sc_init(dir, init, prm_biol, fgs, bboxes, pred = "Cod", save_to_disc = FALSE)
plot_sc_init(df = data2, mult_mum, mult_c)
}

