% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-init.R
\name{sc_init}
\alias{sc_init}
\title{Sanity check initial conditions file}
\usage{
sc_init(dir = getwd(), nc, init, prm_biol, fgs, bboxes, mult_mum, mult_c,
  pred = NULL, no_avail = FALSE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{init}{Character string giving the filename of the initial conditions netcdf
file Usually "init[...].nc".}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{mult_mum}{Numeric vector of multiplication factors applied to the initial
mum values.}

\item{mult_c}{Numeric vector of multiplication factors applied to the initial
C values.}

\item{pred}{Vector of predator acronyms to check. If \code{NULL} (default) all age based
predators are selected.}

\item{no_avail}{Boolean indicating if all availabilities should be set to
1 \code{TRUE} or the actual values from the availability matrix are used
\code{FALSE}. Default is \code{FALSE}.}
}
\value{
Named list with the dataframes as list entry saved as .Rda file.
}
\description{
Sanity check initial conditions file
}
\examples{
dir <- system.file("extdata", "gns", package = "atlantistools")
fgs <- "functionalGroups.csv"
init <- "init_simple_NorthSea.nc"
nc <- "outputNorthSea.nc"
prm_biol <- "NorthSea_biol_fishing.prm"
bboxes <- get_boundary(load_box(dir = dir, bgm = "NorthSea.bgm"))
mult_mum <- seq(0.5, 10, by = 1)
mult_c <- seq(0.5, 10, by = 1)
no_avail <- FALSE
sc_init(dir, nc, init, prm_biol, fgs, bboxes, mult_mum, mult_c)
sc_init(dir, nc, init, prm_biol, fgs, bboxes, mult_mum, mult_c, pred = "COD")
}

