% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc-overlap.R
\name{sc_overlap}
\alias{sc_overlap}
\title{Calculate 3d overlap of predator groups with their prey over time.}
\usage{
sc_overlap(dir = getwd(), nc_gen, prm_biol, bps, fgs, bboxes, out,
  pred = NULL, save_to_disc = FALSE)
}
\arguments{
\item{dir}{Character string giving the path of the Atlantis model folder.
If data is stored in multiple folders (e.g. main model folder and output
folder) you should use 'NULL' as dir.}

\item{nc_gen}{Character string giving the filename of the general netcdf
file Usually "output[...].nc". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string as nc_gen. In addition set dir to 'NULL' in this
case.}

\item{prm_biol}{Character string giving the filename of the biological
parameterfile. Usually "[...]biol_fishing[...].prm". In case you are using
multiple folders for your model files and outputfiles pass the complete
folder/filename string and set dir to 'NULL'.}

\item{bps}{Vector of character strings giving the complete list of epibenthic
functional groups (Only present in the sediment layer). The names have to match
the column 'Name' in the 'functionalGroups.csv' file.}

\item{fgs}{Character string giving the filename of 'functionalGroups.csv'
file. In case you are using multiple folders for your model files and
outputfiles pass the complete folder/filename string as fgs.
In addition set dir to 'NULL' in this case.}

\item{bboxes}{Integer vector giving the box-id of the boundary boxes.}

\item{out}{Character string giving the filename of the *.Rda output
file. In case you are using different folders for your model data
and output files please pass a complete folder/filename string and
set dir to 'NULL'.}

\item{pred}{Vector of predator acronyms to check. If \code{NULL} (default) all age based
predators are selected.}

\item{save_to_disc}{Logical indicating if the resulting list shall be stored
on the hard-disc (\code{TRUE}) or not (\code{FALSE}).}
}
\description{
Calculate 3d overlap of predator groups with their prey over time.
}
\examples{
\dontrun{
dir <- "c:/backup_z/Atlantis_models/Runs/dummy_01_ATLANTIS_NS/"
nc_gen <- "outputNorthSea.nc"
prm_biol <- "NorthSea_biol_fishing.prm"
fgs <- "functionalGroups.csv"
pred <- NULL
bps <- load_bps(dir = dir, fgs = fgs, init = "init_NorthSea.nc")
bboxes <- get_boundary(load_box(dir = dir, bgm = "NorthSea.bgm"))
}
}

