% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-prm.R
\name{scan_prm}
\alias{scan_prm}
\title{Scan character vector for specific flag!}
\usage{
scan_prm(chars, variable)
}
\arguments{
\item{chars}{Vector of character strings}

\item{variable}{Character string giving the flag to search for.}
}
\description{
Scan character vector for specific flag!
}
