% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan-prm.R
\name{scan_prm}
\alias{scan_prm}
\title{Scan character vector for specific flag!}
\usage{
scan_prm(chars, variable, ignore_duplicates = FALSE)
}
\arguments{
\item{chars}{Vector of character strings}

\item{variable}{Character string giving the flag to search for.}

\item{ignore_duplicates}{TRUE to just return first value in case of duplicates, FALSE for error}
}
\description{
Scan character vector for specific flag!
}
